### 2025-04-02 #7989 @adperezm
> [@adperezm](https://github.com/adperezm) - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
>
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

Dear @arfon, thank you for your feedback! We will wait as needed.
