### 2025-06-26 #7989 @adperezm
Hi @finsberg , thank you very much for being the editor for this submission.

After checking the database I think the following reviewers would be reasonable:

[kris-rowe](https://github.com/kris-rowe): Does not specify Python, but spectral element method experience would be relevant.
[baperry2](https://github.com/baperry2): Experience with high performance computing and fluid mechanics would be relevant.
[SoilRos](https://github.com/SoilRos): Has experience with finite element methods and parallel computing as well.

Thanks for the help! Let me know if I should provide more suggestions.
