### 2025-06-27 #7989 @finsberg
Hi @kris-rowe, @baperry2, @SoilRos, @ioannisPApapadopoulos, @danielpeter  👋
Considering your expertise, would any of you be willing to review this submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://github.com/openjournals/joss-papers/blob/joss.07989/joss.07989/10.21105.joss.07989.pdf
Software: https://github.com/ExtremeFLOW/pySEMTools
