### 2025-06-27 #7989 @baperry2
This is a bit out of my wheelhouse and I don't have much bandwidth right now, so I will not be able to review this submission.
