### 2025-06-27 #7989 @SoilRos
@finsberg thanks for the consideration! Unfortunately, I cannot help with the review at this moment. Still, the project sounds interesting, so I wish you @adperezm much success with the paper!
