### 2025-08-08 #8767 @editorialbot
[REVIEW]: PySEMTools: A library for post-processing hexahedral spectral element data.
**Submitting author:** @adperezm (<a href="http://orcid.org/0000-0001-5204-8549">Adalberto Perez</a>)
**Repository:** https://github.com/ExtremeFLOW/pySEMTools
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.0
