### 2025-08-08 #7989 @finsberg
@kris-rowe @nchristensen thank you for your willingness to review this submission. Your feedback is greatly appreciated. The review will continue in the following issue https://github.com/openjournals/joss-reviews/issues/8767
