### 2025-08-08 #8767 @finsberg
@kris-rowe, @nchristensen thanks for your help for reviewing this work! This is where the review happens. I kindly ask you to now formally start the review. Follow the instructions above ☝️ to generate a check box list for yourself here to guide you through the process. Let me know if you have any questions. The reviewers are encouraged to submit issues and pull requests on the software repository. When posting issues on the software repository, I appreciate it if you mention openjournals/joss-reviews/issues/8767 so that a link is created to this thread and I can keep track of what is happening.
