+++ pysemtools/io/mesh_orientation.py
@@ -1,3 +1,58 @@
+##################################################################################################################################################
+##################################################################################################################################################
+"""
+    FUNCTIONS SUMMARY
+        
+    This script contains the following functions:
+    
+    1. generate_data(fname_3d)
+            - Reads mesh & field data and extracts element coordinate arrays.
+           
+    2. calculate_face_differences(assigned_data)
+            - Computes normalized average difference vectors between opposing element faces (r, s, t directions).
+            - First calculates the difference between corresponding points on opposing faces, then takes the average,
+            and finally normalizes the result.
+            
+    3. calculate_face_normals(assigned_data)
+            - Calculates raw face normal vectors for each of the six element faces.
+
+    4. calculate_face_averages(assigned_data)
+            - Computes normalized average difference vectors between opposing element faces (for re-verification).
+            - First calculates the average position of each face, then finds the difference between opposing faces,
+            and finally normalizes the result.
+            
+    5. face_mappings(cal_r, cal_s, cal_t)
+            - Maps local element axes (r, s, t) to global axes (x, y, z) based on dominant difference vector directions.
+            - Ensures unique directional assignment per element.
+            - Identifies the principal flow direction for each element.
+
+    6. compare_mappings(mappings, ref_mapping, method)
+            - Compares a mapping set to a reference mapping and computes match percentage.
+
+    7. compare_methods(mappings_fd, mappings_fa)
+            - Compares face-difference-based mappings to face-average-based mappings and computes match percentage (ideally they should be equal. if not then there is some issue)
+
+    8. align_normals(normals, normals_pair, direction_type)
+            - Ensures that paired face normals are consistently oriented and aligned with a reference. Using rank 0's element 0 as the reference, it compares face normals across all elements
+           and adjusts their orientations as needed.
+
+    9. reduce_face_normals(normals)
+            - After alignment, averages the paired face normals to produce a reduced set of normals per element (reducing 6 face normals to 3 per element i.e., one in each axis).
+
+    10. face_normals_mappings(averaged_normals)
+            - Maps element axes using reduced normals and identifies principal flow directions.
+           
+    11. reorder_assigned_data(assigned_data, mappings_fd, ref_mapping_fd)
+            - Reorders element coordinate data to match a given reference axis mapping. Initially, each element's coordinates are in local (r,s,t) space, which can vary in position within the element.
+            - This function maps them to the global (x,y,z) space using rank 0's element 0 as the reference.
+
+    12. correct_axis_signs(assigned_data_new, extract_reference_r, extract_reference_s, extract_reference_t, r_diff_new, s_diff_new, t_diff_new)
+            - Flips element axes if they are sign-inverted relative to a reference element (taken from rank 0's element 0).
+
+"""
+##################################################################################################################################################
+##################################################################################################################################################
+
 from mpi4py import MPI
 import numpy as np
 from ..datatypes.msh import Mesh
@@ -9,18 +64,17 @@ comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 size = comm.Get_size()
 
-def generate_data(fname_3d: str, fname_out: str):
+def generate_data(fname_3d: str):
     """
     Loads and processes 3D field data.
 
[…]
