### 2025-08-19 e8d5bac
Added some memory limits to the tree search
