### 2025-08-22 bc84c36
Try to make hashtables rank estimate faster
