### 2025-09-25 991f302
README addition: Python 3.13+ has compatibility issues
