### 2025-09-28 12a3e24
Better help line for indexer
