### 2025-09-28 2ef346c
Small indexer imporvements
