### 2025-09-30 0f35ed3
Add mesh periodicity mapping function to support dssum computation
