+++ pysemtools/datatypes/msh_connectivity.py
@@ -19,6 +19,8 @@ from .element_slicing import (
     facet_to_vertex_map,
 )
 import sys
+from typing import Tuple, cast
+import math
 
 __all__ = ['MeshConnectivity']
 
@@ -367,8 +369,485 @@ class MeshConnectivity:
                         local_appearances + global_appearances
                     )
 
+    def validate_rank_size(self, total_elements: int, size: int, num_elements: int):
+        """
+            Validates whether the total number of elements and sections can be evenly distributed among ranks.
+                - Criterion 1: Total number of elements must be evenly divisible across all ranks.
+                - Criterion 2: Each cross-section must be assigned an integer number of ranks.
+            
+            Args:
+                total_elements (int): The total number of elements in the dataset.
+                size (int): The number of ranks available for computation.
+                num_elements (int): The number of elements in each section.
+
+            Returns:
+                bool: True if the distribution is valid, False otherwise.
+            """
+        # Criterion 1:  Total number of elements must be evenly divisible across all ranks.
+        elements_per_rank = total_elements // size
+        if total_elements % size != 0:
+            print(f"Error: Criteria 1 Failed,  {total_elements} % {size} != 0 (elements_per_rank = {elements_per_rank})") if self.rt.comm.rank == 0 else None
+            return False
+
+        # Criterion 2: Each cross-section must be assigned an integer number of ranks.
+        ranks_per_section = num_elements // elements_per_rank
+        if num_elements % elements_per_rank != 0:
+            print(f"Error: Criteria 2 Failed, {num_elements} % {elements_per_rank} != 0 (ranks_per_section = {ranks_per_section})") if self.rt.comm.rank == 0 else None
+            return False
+
+        print(f"Both Criteria Passed: elements_per_rank = {elements_per_rank}, ranks_per_section = {ranks_per_section}") if self.rt.comm.rank == 0 else None
+        return True
+
+    def find_valid_sizes(self, total_elements: int, num_elements: int, min_size=150, max_size=2000):
+        """
+        Finds valid rank sizes for distributing elements evenly across computational processes.
+
+        Args:
+            total_elements (int): Total number of elements.
+            num_elements (int): Number of elements per section.
+            min_size (int, optional): Minimum size of ranks to consider. Defaults to 150.
+            max_size (int, optional): Maximum size of ranks to consider. Defaults to 2000.
+
+        Returns:
+            list: A list of valid sizes for distributing elements.
+        """
+        valid_sizes = []
+
+        for size in range(min_size, max_size + 1):
+            if total_elements % size == 0:
+                elements_per_rank = total_elements // size
+                if num_elements % elements_per_rank == 0:
+                    valid_sizes.append(size)
+
+        if not valid_sizes:
+            print("No valid sizes found. Consider adjusting your min/max limits.")    
+        print("Valid size options:", valid_sizes)
+            
+        return valid_sizes
+
+    def get_periodicity_map(self, msh: Mesh = None, offset_vector: Tuple[int, int, int] = (0, 0, 0), num_elements: int = None, pattern_factor: int = 1):    
+        """
+            Notes: One pattern cross-section is formed by repeating a specific group of cross-sections.
+            The number of cross-sections required to complete one full pattern is referred to as the "pattern_factor".
+
+            Generate a periodicity mapping for the given mesh. Ensures that entities like faces, edges, and vertices on the first cross-section of the mesh are matched
+            and mapped to their corresponding periodic partners on the last cross-section, across MPI ranks. It builds/updates shared maps for rank, element, facet, edge, and vertex.
+
+            Parameters:
+                msh (Mesh): The mesh object containing element and vertex data.
+                offset_vector (Tuple[int, int, int]): The offset to apply to the coordinates for periodicity.
+                num_elements (int): The number of elements in each section.
[…]
