### 2025-10-09 ExtremeFLOW/pySEMTools#215 @nchristensen
Consider making more use of Sphinx autodoc
Regarding the checklist at https://github.com/openjournals/joss-reviews/issues/8767, it seems to me the documentation could be improved by adding more API documentation. ~Since you are already using Sphinx, a relatively straightforward way to do this would be to us Sphinx's `autodoc` and `autosummary` functionality. You might have to move some of the triple quotes of the existing doc-strings to make them render correctly though. See https://www.sphinx-doc.org/en/master/tutorial/automatic-doc-generation.html.~

Didn't see you're already using autodoc. That should make it fairly straightforward. Looking at e.g. the ROM code, a lot of the methods have doc-strings. They would just need to be re-formatted a bit so Sphinx can pick up on them.
