### 2025-10-09 ExtremeFLOW/pySEMTools#216 @nchristensen
Consider adding community guidelines
Regarding the checklist at https://github.com/openjournals/joss-reviews/issues/8767, I suggest adding a paragraph about contributing (Github pull requests?), reporting issues (Github issues?), and seeking support (mailing list? Github issues? email?) somewhere in the README or documentation.
