### 2025-10-19 ExtremeFLOW/pySEMTools#217 @nchristensen
Consider including the list of features in the README
The term "post-processing" encompasses quite a bit and Just from a glance at the README I find it a bit difficult to see what problems this software solves. There is a nice list of features in the paper that you might consider including in the README. https://github.com/openjournals/joss-reviews/issues/8767
