### 2025-10-24 ExtremeFLOW/pySEMTools#218 @nchristensen
Consider running examples as part of unit testing
From brief searching it seems like it should be possible to execute ipynbs as part of the test cases. I´d suggest doing this so contributors are aware of any changes that break the examples. Tag https://github.com/openjournals/joss-reviews/issues/8767
