### 2025-10-27 #8767 @finsberg
@adperezm I've noticed a [number of open issues from reviewers](https://github.com/ExtremeFLOW/pySEMTools/issues) in your repository that are awaiting a response.

Could you please make sure to reply to these as soon as possible? For an open-source project, timely communication with users (especially on issues) is essential. A few weeks without a reply is too long.

If you're unable to address them right now, please let us know so we can coordinate.
