### 2025-11-17 c9ae3df
Added new changes to dssum periodic
