+++ pysemtools/datatypes/msh_connectivity.py
@@ -21,6 +21,7 @@ from .element_slicing import (
 import sys
 from typing import Tuple, cast
 import math
+from mpi4py import MPI
 
 __all__ = ['MeshConnectivity']
 
@@ -427,35 +428,39 @@ class MeshConnectivity:
 
     def get_periodicity_map(self, msh: Mesh = None, offset_vector: Tuple[int, int, int] = (0, 0, 0), num_elements: int = None, pattern_factor: int = 1):    
         """
-            Notes: One pattern cross-section is formed by repeating a specific group of cross-sections.
-            The number of cross-sections required to complete one full pattern is referred to as the "pattern_factor".
+            Generate a periodicity mapping for the given mesh. Identifies entities (vertices, edges, facets) that are periodic partners across all MPI ranks 
+            using a ring-exchange scheme. It builds/updates shared maps for rank, element, facet, edge, and vertex.
 
-            Generate a periodicity mapping for the given mesh. Ensures that entities like faces, edges, and vertices on the first cross-section of the mesh are matched
-            and mapped to their corresponding periodic partners on the last cross-section, across MPI ranks. It builds/updates shared maps for rank, element, facet, edge, and vertex.
+            Notation used: 
+                One "pattern cross-section" is formed by repeating a specific group of cross-sections.
+                The number of cross-sections required to complete one full pattern is referred to as the "pattern_factor".
 
             Parameters:
-                msh (Mesh): The mesh object containing element and vertex data.
-                offset_vector (Tuple[int, int, int]): The offset to apply to the coordinates for periodicity.
-                num_elements (int): The number of elements in each cross-section.
-                pattern_factor (int): The number of cross-sections that form one complete periodic pattern.
+            ----------
+            msh : Mesh
+                The mesh object containing element and vertex data.
+            offset_vector : (Tuple[int, int, int])
+                The offset to apply to the coordinates for periodicity.
+            num_elements : (int) 
+                The number of elements in each cross-section.
+            pattern_factor : (int)
+                The number of cross-sections that form one complete pattern cross-section, must be >= 2.
             
-            Function steps:                
+            This function:                
                 1. Validates the rank size to ensure even distribution of elements.
-                2. Identify ranks that belong to the first and last cross-sections based on the number of elements and pattern factor.
-                3. Based on mesh dimensions, computes local vertices, face centers and edge centers.
-                4. Applies the specified offset to these vertices and centers.
-                5. Exchanges relevant data between matching ranks.
-                6. Updates shared maps for rank, elements, facets, edges, and vertices.
+                2. Determines previous and next ranks for circular communication.
+                3. Based on mesh dim ('msh.gdim'), computes local vertices, edge centers and facet centers.
+                4. Applies the specified offset to generate shifted coordinates for matching.
+                5. Performs an MPI ring exchange so each rank receives and compares shifted vertices from every other rank.
+                6. For matched entities, updates the global shared maps of rank, elements and facets/edges/vertices.
 
             Returns:
-                None 
-                    This function updates the mesh object with periodicity mappings.
+            -------
+            None
+                Updates the mesh objects with periodicity mappings.
         """
-
-        self.total_elements = num_elements * pattern_factor                 # Total elements in the file (e.g., 2688)
-        self.elements_per_rank = self.total_elements // self.rt.comm.size   # No. of elements in each rank (e.g., 2688/192 = 14)
-        self.ranks_per_section = num_elements // self.elements_per_rank     # No. of ranks per each cross-section (e.g., 1344/14 = 96)
-
+        # Validate rank size
+        self.total_elements = num_elements * pattern_factor  # Total elements in the file (e.g., 2688)
[…]
