+++ pysemtools/cli/__init__.py
@@ -1,0 +1,1 @@
+""" Command line interface functionalities available in pySEMTools. """
