+++ doc/pages/summary_api.rst
@@ -1,0 +1,11 @@
+API summary
+-----------
+
+The full API documentation for all the classes and functions is provided below.
+
+.. autosummary::
+   :toctree: generated
+   :recursive:
+
+
+   pysemtools
