+++ pysemtools/datatypes/__init__.py
@@ -1,5 +1,9 @@
-from .msh_connectivity import *
-from .msh import *
-from .field import *
-from .coef import *
-from .msh_partitioning import *
+"""Data types to hold and operate with SEM data."""
+
+from .msh_connectivity import MeshConnectivity
+from .msh import Mesh
+from .field import Field, FieldRegistry
+from .coef import Coef
+from .msh_partitioning import MeshPartitioner
+
+__all__ = ["Coef", "Field", "FieldRegistry", "Mesh", "MeshConnectivity", "MeshPartitioner"]
