+++ pysemtools/cli/__init__.py
@@ -1,1 +1,7 @@
 """ Command line interface functionalities available in pySEMTools. """
+
+from .extract_subdomain import main as extract_subdomain
+from .index_files import main as index_files
+from .visnek import main as visnek
+
+__all__ = ["extract_subdomain", "index_files", "visnek"]
