+++ doc/pages/full_api.rst
@@ -1,16 +1,0 @@
-API
---------
-
-This file contains the documentation for the remaining modules, classes and functions not covered in the USER GUIDE or other sections.
-
-cli
-~~~
-
-.. automodule :: pysemtools.cli
-    :members:
-
-.. autofunction :: pysemtools.cli.extract_subdomain.main
-
-.. autofunction :: pysemtools.cli.index_files.main
-
-.. autofunction :: pysemtools.cli.visnek.main
