+++ doc/pages/api.rst
@@ -1,0 +1,44 @@
+API
+--------
+
+PySEMTools is a collection of modules that together provide a comprehensive toolkit for working with spectral element method (SEM) data.
+
+The presently included submodules are:
+
+.. autosummary::
+    :toctree: generated
+
+    pysemtools.cli
+    pysemtools.comm
+    pysemtools.datatypes
+    pysemtools.interpolation
+
+The compression module inside the package is still considered to be experimental and is prone to change, therefore its documentation is not yet included here.
+
+The full API documentation for all the classes and functions is provided below.
+
+-----------------------
+
+cli
+~~~
+
+.. automodule :: pysemtools.cli
+    :members:
+
+comm
+~~~~
+
+.. automodule :: pysemtools.comm
+    :members:
+
+datatypes
+~~~~~~~~~
+
+.. automodule :: pysemtools.datatypes
+    :members:
+
+interpolation
+~~~~~~~~~~~~~
+
+.. automodule :: pysemtools.interpolation
+    :members:
