### 2025-12-02 ccf8ea4
Added monitoring
