### 2025-12-02 66a3622
Added rom to the API docs
