+++ pysemtools/rom/__init__.py
@@ -1,0 +1,7 @@
+""" Set of tools to perform reduced order modeling tasks. Particularly, proper orthogonal decomposition (POD) """
+
+from .pod import POD
+from .io_help import IoHelp
+from .svd import SVD
+
+__all__ = ["POD", "IoHelp", "SVD"]
