### 2025-12-02 45d6fcc
Added community guidelines and list of features
