### 2025-12-03 4d65b34
Start draft of example checks
