### 2025-12-03 fc67827
Updated interpolation examples
