### 2025-12-03 1647a48
Added corrections to interpolation and rom examples. Also to execution script
