+++ tests/test_examples.py
@@ -108,3 +108,32 @@ def test_interpolation(sem_data_path):
         passed.append(_passed)
 
     assert all(passed)
+
+def test_reduced_order_modelling():
+
+
+    examples_path = "examples/5-reduced_order_modelling/"
+    notebook_files = ["1-POD_from_pointclouds.ipynb", 
+                      "2-POD_fft_from_pointclouds.ipynb"]
+
+    passed = []
+    for notebook in notebook_files:
+        notebook_path = os.path.join(examples_path, notebook)
+
+        print(f"Executing notebook: {notebook_path}")
+        _path=os.path.abspath(notebook_path)
+        _name=os.path.basename(notebook_path)
+        command = f"jupyter nbconvert --to notebook --execute {notebook_path} --output ./executed_{_name}"
+        process = subprocess.run(command, capture_output=True, shell=True)
+        if process.returncode != 0:
+            print("Error executing notebook:")
+            print(process.stdout.decode())
+            print(process.stderr.decode())
+            _passed = False
+        else:
+            print("Notebook executed successfully.")
+            _passed = True
+
+        passed.append(_passed)
+
+    assert all(passed)
