### 2025-12-03 a3b0b28
Added first part of statistics to tests
