### 2025-12-05 14ef006
Added more statistics examples
