### 2025-12-05 d98142e
Finalized examples
