+++ examples/6-statistics/README.md
@@ -3,3 +3,6 @@
 In these examples we show how to use the functionalities of pysemtools to post process statistics.
 
 1. Post processing mean fields from a nek type simulation.
+4. Get turbulent kinetic energy budgets from a pre-computed file with all relevant statistics.
+5. Spatial averaging cases with one clear homogenous direction. Mapping from 3D to 2D files.
+6. How to, from an intial set of snapshots, produce augmented data that can be used to calculate high-order statistics, as in example 4.
