+++ examples/5-reduced_order_modelling/README.md
@@ -5,4 +5,4 @@ In these examples we show how one can use some wrapper functions to perform POD 
 The wrapper functions generally use classes that are part of the rom submodule. If you wish to use them in a different way, you can check the steps followed in the wrappers.
 
 1. POD in physical space.
-2. POD performing FFT in one homogenous direction.
+2. POD performing FFT in one homogenous direction (2.5 is a python script with the same information. To be run in parallel.)
