### 2025-12-05 3d4ff99
Attempt to fix stats example 1
