+++ .github/workflows/check_unit_test.yml
@@ -69,14 +69,13 @@ jobs:
           pip install jupyter
           pip install matplotlib
           pip install pyevtk
-          pip install pyvista
 
           echo "Installing PySEMTools"
           pip install .
           
       - name: Run unit tests
         run: |
-          pytest --json-report -v tests/ | tee unit_tests.txt
+          pytest --json-report -v tests/test_examples.py | tee unit_tests.txt
       
       - name: Check results
         run: |
