### 2025-12-05 480b761
Keep debugging CI - now add some diagnostics.
