### 2025-12-05 d3d588e
Re added examples
