### 2025-12-05 fc4cc64
Added more verbose log
