+++ .github/workflows/check_unit_test.yml
@@ -75,7 +75,7 @@ jobs:
           
       - name: Run unit tests
         run: |
-          pytest --json-report -v tests/test_examples.py | tee unit_tests.txt
+          pytest --json-report -s -v tests/test_examples.py | tee unit_tests.txt
       
       - name: Check results
         run: |
