### 2025-12-05 b1c9b57
Remove budgets from testsuite. To be checked later.
