+++ tests/test_examples.py
@@ -141,7 +141,8 @@ def test_reduced_order_modelling():
 def test_statistics(sem_data_path):
 
     examples_path = "examples/6-statistics/"
-    notebook_files = ["1-post_processing_mean_fields.ipynb", 
+    notebook_files = ["1-post_processing_mean_fields.ipynb",
+                      "4-budgets.ipynb", 
                       "5-homogeneous_direction_averaging.ipynb",
                       "6-Stats_from_fld.ipynb"]
 
