### 2025-12-10 7be762d
Added some classes for educational solver
