+++ examples/3-calculus_in_sem_mesh/solver_2.py
@@ -1,0 +1,171 @@
+# Import required modules
+from mpi4py import MPI #equivalent to the use of MPI_init() in C
+import matplotlib.pyplot as plt
+import numpy as np
+
+# Get mpi info
+comm = MPI.COMM_WORLD
+
+from pysemtools.io.ppymech.neksuite import pynekread
+from pysemtools.datatypes.msh import Mesh
+from pysemtools.datatypes.coef import Coef
+from pysemtools.datatypes.field import FieldRegistry
+from pysemtools.solver.operators import AdvectionOperator, StiffnessOperator
+
+msh = Mesh(comm, create_connectivity=True)
+fld = FieldRegistry(comm)
+pynekread('../data/mixlay0.f00001', comm, data_dtype=np.double, msh=msh, fld=fld)
+coef = Coef(msh, comm, get_area=False)
+from pysemtools.datatypes.msh_connectivity import MeshConnectivity
+msh_conn = MeshConnectivity(comm, msh, rel_tol=1e-5)
+
+
+# ---------------------------------------------------------
+# Parameters
+# ---------------------------------------------------------
+min_dx = np.min(coef.dxdr)
+
+
+dt = min_dx / 1.0 / 10 # time step
+#dt = 0.001
+print(f"Using time step dt = {dt:.3e}")
+kappa = 0.01  # diffusion coefficient
+
+# ---------------------------------------------------------
+# Initial condition
+# ---------------------------------------------------------
+th = np.zeros_like(fld.registry['u'])
+u  = np.ones_like(fld.registry['u'])   # constant velocity in x
+v  = np.zeros_like(fld.registry['u'])  # zero velocity in y
+
+x0, y0 = 10.0, 7.0      # center of the Gaussian
+sigma = 1.0             # width parameter
+
+r2 = (msh.x - x0)**2 + (msh.y - y0)**2
+th = np.exp(-r2 / (2.0 * sigma**2))
+
+# Convective RHS history: rhs_hist[0] = N_adv^n, rhs_hist[1] = N_adv^{n-1}, rhs_hist[2] = N_adv^{n-2}
+rhs_hist = [np.zeros_like(th) for _ in range(3)]
+
+# Initialize operators
+
+advection_op = AdvectionOperator(coef, msh_conn)
+stiffness_op = StiffnessOperator(coef, msh_conn)
+
+# ---------------------------------------------------------
+# Conjugate Gradient solver (matrix-free)
+# ---------------------------------------------------------
+def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
+    """
+    Solve A x = b using Conjugate Gradient.
+    apply_A: function that computes A(x).
+    """
+    if x0 is None:
+        x = np.zeros_like(b)
+    else:
+        x = x0.copy()
+
+    r = b - apply_A(x)
+    p = r.copy()
+    rsold = np.vdot(r, r).real
+    rsold = coef.glsum(rsold, comm=comm)  # global sum
+
+    if rsold == 0.0:
+        return x
+
+    for it in range(1, maxiter + 1):
+        Ap = apply_A(p)
+        pAp = np.vdot(p, Ap).real
+        pAp = coef.glsum(pAp, comm=comm)  # global sum
+        if pAp == 0.0:
+            break
+
+        alpha = rsold / pAp
+        x = x + alpha * p
+        r = r - alpha * Ap
+        rsnew = np.vdot(r, r).real
+        rsnew = coef.glsum(rsnew, comm=comm)  # global sum
+
+        res_check = np.sqrt(rsnew)/coef.glsum(coef.B, comm=comm)
+
+        if res_check < tol:
[…]
