### 2025-12-11 c218e54
Lid solver draft
