+++ examples/3-calculus_in_sem_mesh/solver_3.py
@@ -1,0 +1,290 @@
+# Import required modules
+import sys
+from mpi4py import MPI #equivalent to the use of MPI_init() in C
+import matplotlib.pyplot as plt
+import numpy as np
+
+# Get mpi info
+comm = MPI.COMM_WORLD
+from pysemtools.io.ppymech import pynekwrite, pynekread
+from pysemtools.datatypes import Mesh, Coef, FieldRegistry, MeshConnectivity
+from pysemtools.solver.operators import AdvectionOperator, StiffnessOperator
+
+# -------------------------------------------------------------
+# Helper functions
+# -------------------------------------------------------------
+
+def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
+    """
+    Solve A x = b using Conjugate Gradient.
+    apply_A: function that computes A(x).
+    """
+    if x0 is None:
+        x = np.zeros_like(b)
+    else:
+        x = x0.copy()
+
+    r = b - apply_A(x)
+    p = r.copy()
+    rsold = np.vdot(r, r).real
+    rsold = coef.glsum(rsold, comm=comm)  # global sum
+
+    if rsold == 0.0:
+        return x
+
+    for it in range(1, maxiter + 1):
+        Ap = apply_A(p)
+        pAp = np.vdot(p, Ap).real
+        pAp = coef.glsum(pAp, comm=comm)  # global sum
+        if pAp == 0.0:
+            break
+
+        alpha = rsold / pAp
+        x = x + alpha * p
+        r = r - alpha * Ap
+        rsnew = np.vdot(r, r).real
+        rsnew = coef.glsum(rsnew, comm=comm)  # global sum
+
+        res_check = np.sqrt(rsnew)/coef.glsum(coef.B, comm=comm)
+
+        if res_check < tol:
+            print(f"CG converged in {it} iterations, residual {res_check:.3e}")
+            break
+
+        p = r + (rsnew / rsold) * p
+        rsold = rsnew
+
+    print(f"CG finished in {it} iterations, residual {res_check:.3e}")
+    return x
+
+# -------------------------------------------------------------
+# Read mesh and set up connectivity
+# -------------------------------------------------------------
+msh = Mesh(comm, create_connectivity=False)
+fld = FieldRegistry(comm)
+pynekread('./lid_mesh0.f00000', comm, data_dtype=np.double, msh=msh, fld=fld)
+coef = Coef(msh, comm, get_area=False)
+conn = MeshConnectivity(comm, msh, use_hashtable=True)
+
+# -------------------------------------------------------------
+# Identifying the boundaries and setting up boundary conditions
+# -------------------------------------------------------------
+x = msh.x
+y = msh.y
+x_min, x_max = comm.allreduce(np.min(x), op=MPI.MIN), comm.allreduce(np.max(x), op=MPI.MAX)
+y_min, y_max = comm.allreduce(np.min(y), op=MPI.MIN), comm.allreduce(np.max(y), op=MPI.MAX)
+eps = 1e-8
+
+tol_v = 1e-5
+tol_p = 1e-4
+
+is_left   = np.abs(x - x_min) < eps
+is_right  = np.abs(x - x_max) < eps
+is_bottom = np.abs(y - y_min) < eps
+is_top    = np.abs(y - y_max) < eps
+
+lid_speed = 1.0   # U_lid
+
+boundary_ids = np.zeros_like(msh.x)
+boundary_ids[is_left]   = 1  # left wall
+boundary_ids[is_right]  = 2  # right wall
+boundary_ids[is_bottom] = 3  # bottom wall
[…]
