### 2025-12-11 0fbe505
Refined more the lid solver
