+++ examples/3-calculus_in_sem_mesh/solver_3.py
@@ -9,6 +9,24 @@ comm = MPI.COMM_WORLD
 from pysemtools.io.ppymech import pynekwrite, pynekread
 from pysemtools.datatypes import Mesh, Coef, FieldRegistry, MeshConnectivity
 from pysemtools.solver.operators import AdvectionOperator, StiffnessOperator
+from pysemtools.monitoring import Logger
+
+# -------------------------------------------------------------
+# Simulation parameters
+# -------------------------------------------------------------
+rho = 1.0
+Re = 3000
+nu  = 1/Re   # viscosity
+t_end = 300
+dt  = 0.25e-2    # as you already choose it
+nsteps = int(t_end / dt)
+
+tol_v = 1e-7
+tol_p = 1e-4
+max_iter_v = 100
+max_iter_p = 100
+
+output_interval_t = 1
 
 # -------------------------------------------------------------
 # Helper functions
@@ -45,21 +63,33 @@ def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
         rsnew = np.vdot(r, r).real
         rsnew = coef.glsum(rsnew, comm=comm)  # global sum
 
-        res_check = np.sqrt(rsnew)/coef.glsum(coef.B, comm=comm)
+        #res_check = np.sqrt(rsnew)/coef.glsum(coef.B, comm=comm)
+        res_check = np.sqrt(rsnew)/(msh.glb_nelv*msh.lxyz) # Normalized residual - feels a bit like cheating...
 
         if res_check < tol:
-            print(f"CG converged in {it} iterations, residual {res_check:.3e}")
+            #print(f"CG converged in {it} iterations, residual {res_check:.3e}")
             break
 
         p = r + (rsnew / rsold) * p
         rsold = rsnew
 
-    print(f"CG finished in {it} iterations, residual {res_check:.3e}")
-    return x
+    #print(f"CG finished in {it} iterations, residual {res_check:.3e}")
+    return x, res_check, it
 
 # -------------------------------------------------------------
 # Read mesh and set up connectivity
 # -------------------------------------------------------------
+log = Logger(comm=comm, module_name="LidDrivenCavitySolver")
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Lid Driven Cavity Solver - SEM")
+log.write("info", "---------------------------------------------")
+
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Reading mesh and setting up connectivity")
+log.write("info", "---------------------------------------------")
+
 msh = Mesh(comm, create_connectivity=False)
 fld = FieldRegistry(comm)
 pynekread('./lid_mesh0.f00000', comm, data_dtype=np.double, msh=msh, fld=fld)
@@ -69,14 +99,16 @@ conn = MeshConnectivity(comm, msh, use_hashtable=True)
 # -------------------------------------------------------------
 # Identifying the boundaries and setting up boundary conditions
 # -------------------------------------------------------------
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Setting up boundary conditions")
+log.write("info", "---------------------------------------------")
+
 x = msh.x
 y = msh.y
 x_min, x_max = comm.allreduce(np.min(x), op=MPI.MIN), comm.allreduce(np.max(x), op=MPI.MAX)
 y_min, y_max = comm.allreduce(np.min(y), op=MPI.MIN), comm.allreduce(np.max(y), op=MPI.MAX)
 eps = 1e-8
-
-tol_v = 1e-5
-tol_p = 1e-4
 
[…]
