### 2025-12-11 978fe2a
Removed original transport eq solver
