### 2025-12-11 491d8c1
Use smooth step from neko example
