### 2025-12-15 4880545
Some extra features to finish this side project
