+++ examples/7-solving_NS/1-lid_driven_cavity.py
@@ -23,8 +23,8 @@ nsteps = int(t_end / dt)
 
 tol_v = 1e-7
 tol_p = 1e-4
-max_iter_v = 100
-max_iter_p = 100
+max_iter_v = 1000
+max_iter_p = 1000
 
 output_interval_t = 1
 
@@ -32,10 +32,11 @@ output_interval_t = 1
 # Helper functions
 # -------------------------------------------------------------
 
-def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
+def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200, project=None):
     """
     Solve A x = b using Conjugate Gradient.
-    apply_A: function that computes A(x).
+    apply_A : function that returns A(x)
+    project : optional function q -> q_proj (e.g. zero-mean projection for Neumann Poisson)
     """
     if x0 is None:
         x = np.zeros_like(b)
@@ -43,27 +44,38 @@ def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
         x = x0.copy()
 
     r = b - apply_A(x)
-    p = r.copy()
+    if project is not None:
+        r = project(r)
+
+    # rsold = (r, r) globally
     rsold = np.vdot(r, r).real
-    rsold = coef.glsum(rsold, comm=comm)  # global sum
+    rsold = coef.glsum(rsold, comm=comm)
 
     if rsold == 0.0:
-        return x
+        return x, 0.0, 0
+
+    p = r.copy()
+    N_global = msh.glb_nelv * msh.lxyz
 
     for it in range(1, maxiter + 1):
         Ap = apply_A(p)
         pAp = np.vdot(p, Ap).real
-        pAp = coef.glsum(pAp, comm=comm)  # global sum
+        pAp = coef.glsum(pAp, comm=comm)
         if pAp == 0.0:
             break
 
         alpha = rsold / pAp
         x = x + alpha * p
         r = r - alpha * Ap
-        rsnew = np.vdot(r, r).real
-        rsnew = coef.glsum(rsnew, comm=comm)  # global sum
 
-        res_check = np.sqrt(rsnew)/(msh.glb_nelv*msh.lxyz) # Normalized residual - feels a bit like cheating...
+        if project is not None:
+            r = project(r)
+
+        rsnew = np.vdot(r, r).real
+        rsnew = coef.glsum(rsnew, comm=comm)
+
+        # RMS residual
+        res_check = np.sqrt(rsnew / N_global)
 
         if res_check < tol:
             break
@@ -72,6 +84,115 @@ def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200):
         rsold = rsnew
 
     return x, res_check, it
+
+def pcg_solve(apply_A,
+              b,
+              apply_Minv=None,
+              x0=None,
+              tol=1e-8,
+              maxiter=200,
+              project=None):
+    """
+    Preconditioned Conjugate Gradient:
+        A x = b
+
+    Parameters
+    ----------
+    apply_A : callable
+        y = apply_A(x) = A @ x  (matrix-free operator)
+    apply_Minv : callable or None
+        z = apply_Minv(r) = M^{-1} @ r  (Jacobi, etc).
+        If None, no preconditioning (M = I).
+    project : callable or None
+        q_proj = project(q). Use this to remove nullspace components,
+        e.g. de-mean for Neumann Poisson.
+    """
+
+    # Initial guess
+    if x0 is None:
+        x = np.zeros_like(b)
+    else:
+        x = x0.copy()
+
[…]
