+++ examples/7-solving_NS/1-lid_driven_cavity.py
@@ -9,6 +9,7 @@ comm = MPI.COMM_WORLD
 from pysemtools.io.ppymech import pynekwrite, pynekread
 from pysemtools.datatypes import Mesh, Coef, FieldRegistry, MeshConnectivity
 from pysemtools.solver.operators import AdvectionOperator, StiffnessOperator
+from pysemtools.solver.linear_solver import LinearSolver
 from pysemtools.monitoring import Logger
 
 # -------------------------------------------------------------
@@ -27,168 +28,6 @@ max_iter_v = 1000
 max_iter_p = 1000
 
 output_interval_t = 1
-
-# -------------------------------------------------------------
-# Helper functions
-# -------------------------------------------------------------
-
-def cg_solve(apply_A, b, x0=None, tol=1e-8, maxiter=200, project=None):
-    """
-    Solve A x = b using Conjugate Gradient.
-    apply_A : function that returns A(x)
-    project : optional function q -> q_proj (e.g. zero-mean projection for Neumann Poisson)
-    """
-    if x0 is None:
-        x = np.zeros_like(b)
-    else:
-        x = x0.copy()
-
-    r = b - apply_A(x)
-    if project is not None:
-        r = project(r)
-
-    # rsold = (r, r) globally
-    rsold = np.vdot(r, r).real
-    rsold = coef.glsum(rsold, comm=comm)
-
-    if rsold == 0.0:
-        return x, 0.0, 0
-
-    p = r.copy()
-    N_global = msh.glb_nelv * msh.lxyz
-
-    for it in range(1, maxiter + 1):
-        Ap = apply_A(p)
-        pAp = np.vdot(p, Ap).real
-        pAp = coef.glsum(pAp, comm=comm)
-        if pAp == 0.0:
-            break
-
-        alpha = rsold / pAp
-        x = x + alpha * p
-        r = r - alpha * Ap
-
-        if project is not None:
-            r = project(r)
-
-        rsnew = np.vdot(r, r).real
-        rsnew = coef.glsum(rsnew, comm=comm)
-
-        # RMS residual
-        res_check = np.sqrt(rsnew / N_global)
-
-        if res_check < tol:
-            break
-
-        p = r + (rsnew / rsold) * p
-        rsold = rsnew
-
-    return x, res_check, it
-
-def pcg_solve(apply_A,
-              b,
-              apply_Minv=None,
-              x0=None,
-              tol=1e-8,
-              maxiter=200,
-              project=None):
-    """
-    Preconditioned Conjugate Gradient:
-        A x = b
-
-    Parameters
-    ----------
-    apply_A : callable
-        y = apply_A(x) = A @ x  (matrix-free operator)
-    apply_Minv : callable or None
-        z = apply_Minv(r) = M^{-1} @ r  (Jacobi, etc).
-        If None, no preconditioning (M = I).
-    project : callable or None
-        q_proj = project(q). Use this to remove nullspace components,
-        e.g. de-mean for Neumann Poisson.
-    """
-
-    # Initial guess
-    if x0 is None:
-        x = np.zeros_like(b)
-    else:
-        x = x0.copy()
-
-    # r0 = b - A x0
-    r = b - apply_A(x)
-    if project is not None:
-        r = project(r)
-
-    # Preconditioned residual z0 = M^{-1} r0
-    if apply_Minv is None:
-        z = r.copy()
-    else:
-        z = apply_Minv(r)
-
-    # Dot products (global)
-    rz_old = np.vdot(r, z).real
-    rz_old = coef.glsum(rz_old, comm=comm)
-
[…]
