### 2025-12-16 6347684
Added a simplified cylinder example
