+++ examples/7-solving_NS/2-2d_cylinder.py
@@ -1,0 +1,319 @@
+# Import required modules
+import sys
+from mpi4py import MPI #equivalent to the use of MPI_init() in C
+import matplotlib.pyplot as plt
+import numpy as np
+
+# Get mpi info
+comm = MPI.COMM_WORLD
+from pysemtools.io.ppymech import pynekwrite, pynekread
+from pysemtools.datatypes import Mesh, Coef, FieldRegistry, MeshConnectivity
+from pysemtools.solver.operators import AdvectionOperator, StiffnessOperator
+from pysemtools.solver.linear_solver import LinearSolver
+from pysemtools.monitoring import Logger
+
+# -------------------------------------------------------------
+# Simulation parameters
+# -------------------------------------------------------------
+rho = 1.0
+Re = 200
+nu  = 1/Re   # viscosity
+t_end = 200
+dt  = 5.0e-3    # as you already choose it
+nsteps = int(t_end / dt)
+
+tol_v = 1e-7
+tol_p = 1e-4
+max_iter_v = 1000
+max_iter_p = 1000
+
+output_interval_t = 1
+
+def de_mean_pressure_rhs(q):
+    q_mean = coef.glsum(q, comm=comm) / (msh.glb_nelv*msh.lxyz)
+    return q - q_mean
+
+# -------------------------------------------------------------
+# Read mesh and set up connectivity
+# -------------------------------------------------------------
+log = Logger(comm=comm, module_name="CylinderFlowSolver")
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Cylinder Flow Solver - SEM")
+log.write("info", "---------------------------------------------")
+
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Reading mesh and setting up connectivity")
+log.write("info", "---------------------------------------------")
+
+msh = Mesh(comm, create_connectivity=False)
+fld = FieldRegistry(comm)
+pynekread('./../data/cylinder_mesh0.f00000', comm, data_dtype=np.double, msh=msh, fld=fld)
+coef = Coef(msh, comm, get_area=False)
+conn = MeshConnectivity(comm, msh, use_hashtable=True, max_elem_per_vertex=5, coef=coef)
+
+# -------------------------------------------------------------
+# Identifying the boundaries and setting up boundary conditions
+# -------------------------------------------------------------
+
+log.write("info", "---------------------------------------------")
+log.write("info", "Setting up boundary conditions")
+log.write("info", "---------------------------------------------")
+
+x = msh.x
+y = msh.y
+x_min, x_max = comm.allreduce(np.min(x), op=MPI.MIN), comm.allreduce(np.max(x), op=MPI.MAX)
+y_min, y_max = comm.allreduce(np.min(y), op=MPI.MIN), comm.allreduce(np.max(y), op=MPI.MAX)
+eps = 1e-8
+
+is_left   = np.abs(x - x_min) < eps
+is_right  = np.abs(x - x_max) < eps
+is_bottom = np.abs(y - y_min) < eps
+is_top    = np.abs(y - y_max) < eps
+is_cylinder = fld.registry["s0"]==1  # assuming s0 = 1 marks the cylinder
+
+boundary_ids = np.zeros_like(msh.x)
+boundary_ids[is_left]   = 1  # left wall
+boundary_ids[is_right]  = 2  # right wall
+boundary_ids[is_bottom] = 3  # bottom wall
+boundary_ids[is_top]    = 4  # top wall
+boundary_ids[is_cylinder] = 5  # cylinder wall
+
+# Write out the boundaries to check
[…]
