### 2026-01-01 #8767 @finsberg
@nchristensen and @kris-rowe, could you please check the lates changes from @adperezm and see if you are satisfied. We will need all checkboxes in your lists to be ticked off by both of you before we can proceed. Thanks again for the help with reviewing this.
