### 2026-01-05 #8767 @nchristensen
Possibly not actionable, but it would be useful if JOSS could configure Pandoc could include the URL in the citations if a DOI is not included. In the proof, the Nek5000 website citation doesn't show the URL.
