### 2026-01-05 d20300d
Add url entry to Nek5000 bib item to make sure url renders in pdf
