+++ paper.bib
@@ -16,6 +16,7 @@ abstract = {A spectral element method that combines the generality of the finite
   note   = {\url{http://nek5000.mcs.anl.gov}},
   title  = {{Nek5000} {w}eb page},
   year   = {2008},
+  url  = {http://nek5000.mcs.anl.gov},
 }
 
 
