### 2026-01-05 e706422
Example script for hpod
