+++ scripts/2-rom/5-hpod/sem_run.py
@@ -1,0 +1,184 @@
+# Import general modules
+import sys
+import os
+os.environ["OMP_NUM_THREADS"] = "1"
+os.environ["OPENBLAS_NUM_THREADS"] = "1"
+os.environ["MKL_NUM_THREADS"] = "1"
+os.environ["VECLIB_MAXIMUM_THREADS"] = "1"
+os.environ["NUMEXPR_NUM_THREADS"] = "1"
+
+import json
+import numpy as np
+import matplotlib.pyplot as plt
+
+# Import MPI
+from mpi4py import MPI #equivalent to the use of MPI_init() in C
+
+# Split communicator for MPI - MPMD
+worldcomm = MPI.COMM_WORLD
+worldrank = worldcomm.Get_rank()
+worldsize = worldcomm.Get_size()
+col = 1
+comm = worldcomm.Split(col,worldrank)
+rank = comm.Get_rank()
+size = comm.Get_size()
+
+#=========================================
+# Define inputs
+#=========================================
+
+# Open input file to see path
+f = open ("inputs.json", "r")
+params_file = json.loads(f.read())
+f.close()
+
+# Read the POD inputs
+pod_number_of_snapshots = params_file["number_of_snapshots"]
+pod_fields = params_file["fields"]
+number_of_pod_fields = len(pod_fields)
+pod_batch_size = params_file["batch_size"]
+pod_keep_modes = params_file["keep_modes"]
+pod_write_modes = params_file["write_modes"]
+dtype_string = params_file.get("dtype", "double")
+backend = params_file.get("backend", "numpy")
+if dtype_string == "single":
+    dtype = np.float32
+else:
+    dtype = np.float64
+
+# Import IO helper functions
+from pysemtools.io.utils import get_fld_from_ndarray, IoPathData
+# Import modules for reading and writing
+from pysemtools.io.ppymech.neksuite import pynekread, pynekwrite
+# Import the data types
+from pysemtools.datatypes.msh import Mesh
+from pysemtools.datatypes.coef import Coef
+from pysemtools.datatypes.field import Field, FieldRegistry
+from pysemtools.datatypes.utils import create_hexadata_from_msh_fld
+# Import types asociated with POD
+from pysemtools.rom.pod import POD
+from pysemtools.rom.io_help import IoHelp
+
+# Start time
+start_time = MPI.Wtime()
+
+#=========================================
+# Get the mesh
+#=========================================
+
+# Read the data paths from the input file
+mesh_data = IoPathData(params_file["IO"]["mesh_data"])
+field_data = IoPathData(params_file["IO"]["field_data"])
+
+#=========================================
+# Initialize the POD
+#=========================================
+
+# Instance the POD object
+pod = POD(comm, number_of_modes_to_update = pod_keep_modes, global_updates = True, auto_expand = False, bckend = backend)
+
+# Initialize the mesh file
+path     = mesh_data.dataPath
+casename = mesh_data.casename
+index    = mesh_data.index
+fname    = path+casename+'0.f'+str(index).zfill(5)
+msh = Mesh(comm)
+pynekread(fname, comm, data_dtype=dtype, msh = msh)
+
+# Initialize coef to get the mass matrix
+coef = Coef(msh, comm)
+bm = coef.B
+
+# Instance io helper that will serve as buffer for the snapshots
+ioh = IoHelp(comm, number_of_fields = number_of_pod_fields, batch_size = pod_batch_size, field_size = bm.size, field_data_type=dtype, mass_matrix_data_type=dtype)
+
+# Put the mass matrix in the appropiate format (long 1d array)
+mass_list = []
+for i in range(0, number_of_pod_fields):
+    mass_list.append(np.copy(np.sqrt(bm)))
+ioh.copy_fieldlist_to_xi(mass_list)
[…]
