### 2026-01-15 9a60aad
Some changes to the data types logging
