+++ pysemtools/interpolation/interpolator.py
@@ -279,8 +279,8 @@ class Interpolator:
         local_data_structure="kdtree",
     ):
         self.log = Logger(comm=comm, module_name="Interpolator")
-        self.log.write("info", "Initializing Interpolator object")
-        self.log.tic()
+        self.log.write("debug", "Initializing Interpolator object")
+        #self.log.tic()
 
         # Instance communication object
         self.rt = Router(comm)
@@ -306,7 +306,7 @@ class Interpolator:
 
         # Determine which point interpolator to use
         self.log.write(
-            "info",
+            "debug",
             "Initializing point interpolator: {}".format(point_interpolator_type),
         )
         self.ei = get_point_interpolator(
@@ -318,7 +318,7 @@ class Interpolator:
         )
 
         # Determine which buffer to use
-        self.log.write("info", "Allocating buffers in point interpolator")
+        self.log.write("debug", "Allocating buffers in point interpolator")
         self.r = self.ei.alloc_result_buffer(dtype="double")
         self.s = self.ei.alloc_result_buffer(dtype="double")
         self.t = self.ei.alloc_result_buffer(dtype="double")
@@ -331,7 +331,7 @@ class Interpolator:
         except AttributeError:
             dev = "cpu"
 
-        self.log.write("info", f"Using device: {dev}")
+        self.log.write("debug", f"Using device: {dev}")
 
         self.progress_bar = progress_bar
 
@@ -378,8 +378,8 @@ class Interpolator:
         self.search_radious = None
         self.bin_to_rank_map = None
 
-        self.log.write("info", "Interpolator initialized")
-        self.log.toc()
+        self.log.write("debug", "Interpolator initialized")
+        #self.log.toc("Interpolator initialized")
 
     def set_up_global_tree(
         self,
@@ -391,12 +391,12 @@ class Interpolator:
 
         if find_points_comm_pattern == "broadcast":
             self.log.write(
-                "info", "Communication pattern selected does not need global tree"
+                "debug", "Communication pattern selected does not need global tree"
             )
 
         elif (find_points_comm_pattern == "point_to_point") or (find_points_comm_pattern == "collective") or (find_points_comm_pattern == "rma"):
             self.global_tree_type = global_tree_type
-            self.log.write("info", f"Using global_tree of type: {global_tree_type}")
+            self.log.write("debug", f"Using global_tree of type: {global_tree_type}")
             if global_tree_type == "rank_bbox":
                 self.set_up_global_tree_rank_bbox_(
                     comm, global_tree_nbins=global_tree_nbins
@@ -408,12 +408,12 @@ class Interpolator:
 
     def set_up_global_tree_rank_bbox_(self, comm, global_tree_nbins=None):
 
-        self.log.tic()
+        #self.log.tic()
 
         size = comm.Get_size()
 
         # Find the bounding box of the rank to create a global but "sparse" kdtree
-        self.log.write("info", "Finding bounding boxes for each rank")
+        self.log.write("debug", "Finding bounding boxes for each rank")
         rank_bbox = np.zeros((1, 6), dtype=INTERPOLATION_DTYPE)
         rank_bbox[0, 0] = np.min(self.x)
         rank_bbox[0, 1] = np.max(self.x)
@@ -449,11 +449,11 @@ class Interpolator:
         bbox_centroid[:, 2] = self.global_bbox[:, 4] + bbox_dist[:, 2] / 2
 
         # Create a tree with the rank centroids
-        self.log.write("info", "Creating global KD tree with rank centroids")
+        self.log.write("debug", "Creating global KD tree with rank centroids")
         self.global_tree = KDTree(bbox_centroid)
[…]
