### 2026-01-15 9e57d75
Finalized logging enhancements on the content level
