### 2026-01-21 a92a89d
fix to make find_points_tol truly optional
