### 2026-02-06 fb22ca2
Initial attempt at compute_and_write_additional_sstat_fields
