+++ pysemtools/postprocessing/statistics/passive_scalar_budgets.py
@@ -1,0 +1,775 @@
+###########################################################################################
+###########################################################################################
+###########################################################################################
+###########################################################################################
+## Preliminary functions to make life easier.
+###########################################################################################
+###########################################################################################
+#%% generic function to compute the gradient of a scalar field
+def compute_scalar_first_derivative(comm, msh, coef, scalar, scalar_deriv):
+    if msh.gdim == 3:
+        scalar_deriv.c1 = coef.dudxyz(scalar, coef.drdx, coef.dsdx, coef.dtdx)
+        scalar_deriv.c2 = coef.dudxyz(scalar, coef.drdy, coef.dsdy, coef.dtdy)
+        scalar_deriv.c3 = coef.dudxyz(scalar, coef.drdz, coef.dsdz, coef.dtdz)
+    elif msh.gdim == 2:
+        scalar_deriv.c1 = coef.dudxyz(scalar, coef.drdx, coef.dsdx, coef.dtdx)
+        scalar_deriv.c2 = coef.dudxyz(scalar, coef.drdy, coef.dsdy, coef.dtdy)
+        scalar_deriv.c3 = 0.0 * scalar_deriv.c2
+    else:
+        import sys
+        sys.exit("supports either 2D or 3D data")
+###########################################################################################
+###########################################################################################
+
+
+#%% generic function to compute the diagonal second derivatives of a scalar field from its gradient
+def compute_scalar_second_derivative(comm, msh, coef, scalar_deriv, scalar_deriv2):
+    if msh.gdim == 3:
+        scalar_deriv2.c1 = coef.dudxyz(scalar_deriv.c1, coef.drdx, coef.dsdx, coef.dtdx)
+        scalar_deriv2.c2 = coef.dudxyz(scalar_deriv.c2, coef.drdy, coef.dsdy, coef.dtdy)
+        scalar_deriv2.c3 = coef.dudxyz(scalar_deriv.c3, coef.drdz, coef.dsdz, coef.dtdz)
+    elif msh.gdim == 2:
+        scalar_deriv2.c1 = coef.dudxyz(scalar_deriv.c1, coef.drdx, coef.dsdx, coef.dtdx)
+        scalar_deriv2.c2 = coef.dudxyz(scalar_deriv.c2, coef.drdy, coef.dsdy, coef.dtdy)
+        scalar_deriv2.c3 = 0.0 * scalar_deriv2.c3
+    else:
+        import sys
+        sys.exit("supports either 2D or 3D data")
+###########################################################################################
+###########################################################################################
+
+
+#%% generic function to write a 9 component field with input as 3 vectors of 3 components each
+def write_file_9c(comm, msh, dU_dxi, dV_dxi, dW_dxi, fname_gradU, if_write_mesh):
+    from pysemtools.datatypes.field import FieldRegistry
+    from pysemtools.io.ppymech.neksuite import pynekwrite
+    import numpy as np
+
+    gradU = FieldRegistry(comm)
+
+    gradU.add_field(comm, field_name="c1", field=dU_dxi.c1, dtype=np.single)
+    gradU.add_field(comm, field_name="c2", field=dU_dxi.c2, dtype=np.single)
+    gradU.add_field(comm, field_name="c3", field=dU_dxi.c3, dtype=np.single)
+    gradU.add_field(comm, field_name="c4", field=dV_dxi.c1, dtype=np.single)
+    gradU.add_field(comm, field_name="c5", field=dV_dxi.c2, dtype=np.single)
+    gradU.add_field(comm, field_name="c6", field=dV_dxi.c3, dtype=np.single)
+    gradU.add_field(comm, field_name="c7", field=dW_dxi.c1, dtype=np.single)
+    gradU.add_field(comm, field_name="c8", field=dW_dxi.c2, dtype=np.single)
[…]
