### 2026-02-06 c5af29b
Add interpolate_all_stat_and_sstat_fields_onto_points
