### 2026-02-12 6bc67bf
working on passive scalar budgets
