+++ pysemtools/postprocessing/statistics/passive_scalar_budgets_interpolatedPoints_notMPI.py
@@ -5,12 +5,13 @@
 #####################################################################################
 # function to read and store the interpolated fields as structured and averaged fields
 #####################################################################################
-def read_interpolated_stat_hdf5_fields( 
+def read_interpolated_scalar_stat_hdf5_fields( 
         path_to_files , 
         Reynolds_number , 
+        Prandtl_number ,
         If_average , If_convert_to_single , 
         Nstruct , av_func , 
-        output_fname='averaged_and_renamed_interpolated_fields.hdf5'
+        output_fname='averaged_and_renamed_interpolated_scalar_fields.hdf5'
         ):
     
     #%%
@@ -19,9 +20,25 @@ def read_interpolated_stat_hdf5_fields(
     import time
 
     #%%
+    # Generic function to load field by number
+    def load_field(num, prefix='interpolated_scalar_fields'):
+        """Load HDF5 field from file by number.
+        
+        Parameters:
+        - num: field number (int)
+        - prefix: filename prefix (default: 'interpolated_scalar_fields')
+        
+        Returns:
+        - Flattened numpy array of field data
+        """
+        filename = path_to_files + '/' + f'{prefix}{num:05d}.hdf5'
+        with h5py.File(filename, 'r') as f:
+            return np.array(f['/field_0']).flatten()
+
+    #%%
     print('reading xyz coordinates...')
     start_time = time.time()
-    with h5py.File(path_to_files+'/'+'coordinates_interpolated_fields.hdf5', 'r') as f:
+    with h5py.File(path_to_files+'/'+'coordinates_interpolated_scalar_fields.hdf5', 'r') as f:
         XYZ_vec = np.array(f['/xyz']).T  # Transpose to match MATLAB's permute([2,1])
     XYZ_vec = XYZ_vec.T 
 
@@ -29,117 +46,43 @@ def read_interpolated_stat_hdf5_fields(
     print(f'Done in {time.time() - start_time:.2f} seconds.')
 
     #%%
-    print('reading the 44 statistics fields...')
+    print('reading the 42 statistics fields...')
     start_time = time.time()
-    UVW_vec = np.zeros((Npts, 3))
-    for i in range(1, 4):
-        filename = path_to_files+'/'+f'interpolated_fields0000{i}.hdf5'
-        with h5py.File(filename, 'r') as f:
-            UVW_vec[:, i-1] = np.array(f['/field_0']).flatten()
-
-    with h5py.File(path_to_files+'/'+'interpolated_fields00004.hdf5', 'r') as f:
-        P_vec = np.array(f['/field_0']).flatten()
-
-    with h5py.File(path_to_files+'/'+'interpolated_fields00005.hdf5', 'r') as f:
-        P2_vec = np.array(f['/field_0']).flatten()
-
-    UiUj_vec = np.zeros((Npts, 6))
-    for i in range(1, 7):
-        tmp = str(5 + i)
-        fnum = ('000000'[:5 - len(tmp)] + tmp)
-        filename = path_to_files+'/'+f'interpolated_fields{fnum}.hdf5'
-        with h5py.File(filename, 'r') as f:
-            UiUj_vec[:, i-1] = np.array(f['/field_0']).flatten()
-
-    UiUjUk_vec = np.zeros((Npts, 10))
-    for i in range(1, 11):
-        tmp = str(11 + i)
-        fnum = ('000000'[:5 - len(tmp)] + tmp)
-        filename = path_to_files+'/'+f'interpolated_fields{fnum}.hdf5'
-        with h5py.File(filename, 'r') as f:
-            UiUjUk_vec[:, i-1] = np.array(f['/field_0']).flatten()
-
[…]
