+++ pysemtools/postprocessing/statistics/passive_scalar_budgets.py
@@ -5,7 +5,7 @@
 ## Preliminary functions to make life easier.
 ###########################################################################################
 ###########################################################################################
-#%% generic function to compute the gradient of a scalar field
+# %% generic function to compute the gradient of a scalar field
 def compute_scalar_first_derivative(comm, msh, coef, scalar, scalar_deriv):
     if msh.gdim == 3:
         scalar_deriv.c1 = coef.dudxyz(scalar, coef.drdx, coef.dsdx, coef.dtdx)
@@ -17,12 +17,15 @@ def compute_scalar_first_derivative(comm, msh, coef, scalar, scalar_deriv):
         scalar_deriv.c3 = 0.0 * scalar_deriv.c2
     else:
         import sys
+
         sys.exit("supports either 2D or 3D data")
+
+
 ###########################################################################################
 ###########################################################################################
 
 
-#%% generic function to compute the diagonal second derivatives of a scalar field from its gradient
+# %% generic function to compute the diagonal second derivatives of a scalar field from its gradient
 def compute_scalar_second_derivative(comm, msh, coef, scalar_deriv, scalar_deriv2):
     if msh.gdim == 3:
         scalar_deriv2.c1 = coef.dudxyz(scalar_deriv.c1, coef.drdx, coef.dsdx, coef.dtdx)
@@ -34,12 +37,15 @@ def compute_scalar_second_derivative(comm, msh, coef, scalar_deriv, scalar_deriv
         scalar_deriv2.c3 = 0.0 * scalar_deriv2.c3
     else:
         import sys
+
         sys.exit("supports either 2D or 3D data")
+
+
 ###########################################################################################
 ###########################################################################################
 
 
-#%% generic function to write a 9 component field with input as 3 vectors of 3 components each
+# %% generic function to write a 9 component field with input as 3 vectors of 3 components each
 def write_file_9c(comm, msh, dU_dxi, dV_dxi, dW_dxi, fname_gradU, if_write_mesh):
     from pysemtools.datatypes.field import FieldRegistry
     from pysemtools.io.ppymech.neksuite import pynekwrite
@@ -60,11 +66,13 @@ def write_file_9c(comm, msh, dU_dxi, dV_dxi, dW_dxi, fname_gradU, if_write_mesh)
     pynekwrite(fname_gradU, comm, msh=msh, fld=gradU, wdsz=4, write_mesh=if_write_mesh)
 
     gradU.clear()
+
+
 ###########################################################################################
 ###########################################################################################
 
 
-#%% generic function to write a 6-component field with input as 2 vectors of 3 components each
+# %% generic function to write a 6-component field with input as 2 vectors of 3 components each
 def write_file_6c(comm, msh, dU_dxi, dV_dxi, fname_gradU, if_write_mesh):
     from pysemtools.datatypes.field import FieldRegistry
     from pysemtools.io.ppymech.neksuite import pynekwrite
@@ -82,11 +90,13 @@ def write_file_6c(comm, msh, dU_dxi, dV_dxi, fname_gradU, if_write_mesh):
     pynekwrite(fname_gradU, comm, msh=msh, fld=gradU, wdsz=4, write_mesh=if_write_mesh)
 
     gradU.clear()
+
+
 ###########################################################################################
 ###########################################################################################
 
 
-#%% generic function to write a 3-component (vector) field
+# %% generic function to write a 3-component (vector) field
[…]
