+++ pysemtools/postprocessing/statistics/passive_scalar_budgets_interpolatedPoints_notMPI.py
@@ -1,54 +1,57 @@
-
-#%%
+# %%
 #####################################################################################
 #####################################################################################
 #####################################################################################
 # function to read and store the interpolated fields as structured and averaged fields
 #####################################################################################
-def read_interpolated_scalar_stat_hdf5_fields( 
-        path_to_files , 
-        Reynolds_number , 
-        Prandtl_number ,
-        If_average , If_convert_to_single , 
-        Nstruct , av_func , 
-        output_fname='averaged_and_renamed_interpolated_scalar_fields.hdf5'
-        ):
-    
-    #%%
+def read_interpolated_scalar_stat_hdf5_fields(
+    path_to_files,
+    Reynolds_number,
+    Prandtl_number,
+    If_average,
+    If_convert_to_single,
+    Nstruct,
+    av_func,
+    output_fname="averaged_and_renamed_interpolated_scalar_fields.hdf5",
+):
+
+    # %%
     import h5py
     import numpy as np
     import time
 
-    #%%
+    # %%
     # Generic function to load field by number
-    def load_field(num, prefix='interpolated_scalar_fields'):
+    def load_field(num, prefix="interpolated_scalar_fields"):
         """Load HDF5 field from file by number.
-        
+
         Parameters:
         - num: field number (int)
         - prefix: filename prefix (default: 'interpolated_scalar_fields')
-        
+
         Returns:
         - Flattened numpy array of field data
         """
-        filename = path_to_files + '/' + f'{prefix}{num:05d}.hdf5'
-        with h5py.File(filename, 'r') as f:
-            return np.array(f['/field_0']).flatten()
+        filename = path_to_files + "/" + f"{prefix}{num:05d}.hdf5"
+        with h5py.File(filename, "r") as f:
+            return np.array(f["/field_0"]).flatten()
 
-    #%%
-    print('reading xyz coordinates...')
+    # %%
+    print("reading xyz coordinates...")
     start_time = time.time()
-    with h5py.File(path_to_files+'/'+'coordinates_interpolated_scalar_fields.hdf5', 'r') as f:
-        XYZ_vec = np.array(f['/xyz']).T  # Transpose to match MATLAB's permute([2,1])
-    XYZ_vec = XYZ_vec.T 
+    with h5py.File(
+        path_to_files + "/" + "coordinates_interpolated_scalar_fields.hdf5", "r"
+    ) as f:
+        XYZ_vec = np.array(f["/xyz"]).T  # Transpose to match MATLAB's permute([2,1])
+    XYZ_vec = XYZ_vec.T
 
     Npts = XYZ_vec.shape[0]
-    print(f'Done in {time.time() - start_time:.2f} seconds.')
+    print(f"Done in {time.time() - start_time:.2f} seconds.")
 
-    #%%
-    print('reading the 42 statistics fields...')
+    # %%
+    print("reading the 42 statistics fields...")
     start_time = time.time()
-    
+
     S_vec = load_field(0)
     UiS_vec = np.column_stack([load_field(i) for i in range(1, 4)])
     S2_vec = load_field(4)
@@ -61,243 +64,648 @@ def read_interpolated_scalar_stat_hdf5_fields(
     UidSdxj_vec = np.column_stack([load_field(i) for i in range(20, 29)])
     SdUidxj_vec = np.column_stack([load_field(i) for i in range(29, 38)])
     SDiss_vec = np.column_stack([load_field(i) for i in range(38, 42)])
-    
-    print(f'Done in {time.time() - start_time:.2f} seconds.')
 
-    #%% make 1D arrays 2D
-    S_vec  = S_vec[:, np.newaxis]
[…]
