### 2026-02-20 13cc15c
Update pyproject.toml with dependencies as add optional depenencies
