### 2026-02-20 36aafd9
Add all dependencies in pyproject.toml
