+++ requirements.txt
@@ -1,3 +1,0 @@
-numpy
-scipy
-pre-commit
