### 2026-02-20 ExtremeFLOW/pySEMTools#235 @finsberg
[JOSS REVIEW] Add package to pypi
You should simplify the installation instructions by adding your package to pypi. Requiring that users clone your repository makes the package challenging to use and in particular very hard to add as a dependency in another library. I added an example on how to write different optional depencies in https://github.com/ExtremeFLOW/pySEMTools/pull/234

- You can read more about writing you `pyproject.toml` here: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
- You can read more about publishing to pypi here: https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/

This is related to the ongoing review at https://github.com/openjournals/joss-reviews/issues/8767
