### 2026-02-20 ExtremeFLOW/pySEMTools#236 @finsberg
[JOSS REIVEW] Fix version in `pyproject.toml`
The version in your `pyproject.toml` does not match the version in your latest release.

This is related to the ongoing review at https://github.com/openjournals/joss-reviews/issues/8767
