### 2026-02-20 ExtremeFLOW/pySEMTools#237 @finsberg
[JOSS REVIEW] Add better instructions on how to run your tests
You write in you README file that to run the tests you only need to install `pytest`, but it will not work without installed additional dependencies. This is in part addressed in https://github.com/ExtremeFLOW/pySEMTools/pull/234, but the instructions in the README should be updated.

This is related to the ongoing review at https://github.com/openjournals/joss-reviews/issues/8767
