### 2026-02-20 ExtremeFLOW/pySEMTools#238 @finsberg
[JOSS REVIEW] Contribution of authorship
Here is a summary showing the contribution to the project from different users (using `git shortlog --summary`):

```
   524  Adalberto
    11  Adalberto Perez
     1  Henrik Finsberg
     5  Nicholas Christensen
     4  Nick Morse
     1  Philipp Schlatter
     2  Shiyu-Sandy-Du
     9  Siavash Toosi
     2  Tim Felle Olsen
    13  Timofey Mukha
     1  Udaya Bhaskar Putta
     3  Victor Baconnet
    14  adalberto
   439  adperezm
     5  iz33agoh
```

It is quite clear the Adalberto Perez is the main contributor, but can you please explain the contribution of the other authors on the paper?

This is related to the ongoing review at https://github.com/openjournals/joss-reviews/issues/8767
