### 2026-02-23 f333e62
First draft of scalar stats complete
