+++ pysemtools/postprocessing/statistics/passive_scalar_budgets.py
@@ -271,10 +271,33 @@ def compute_and_write_additional_sstat_fields(
     file_keys_UiS = ["vel_0", "vel_1", "vel_2"]
     file_keys_SS = ["temp"]
     #                   "USS"      "VSS"     "WSS"
-    file_keys_UjSS = ["scal_2", "scal_3", "scal_4"]
+    file_keys_UiSS = ["scal_2", "scal_3", "scal_4"]
     #                   "UUS"     "VVS"     "WWS"     "UVS"     "UWS"     "VWS"
     file_keys_UiUjS = ["scal_5", "scal_6", "scal_7", "scal_8", "scal_9", "scal_10"]
     file_keys_PS = ["scal_11"]
+
+    file_keys_UidSdxj = [
+        "scal_15",
+        "scal_16",
+        "scal_17",
+        "scal_18",
+        "scal_19",
+        "scal_20",
+        "scal_21",
+        "scal_22",
+        "scal_23",
+    ]
+    file_keys_SdUidxj = [
+        "scal_24",
+        "scal_25",
+        "scal_26",
+        "scal_27",
+        "scal_28",
+        "scal_29",
+        "scal_30",
+        "scal_31",
+        "scal_32",
+    ]
 
     ###########################################################################################
     # S first and second derivatives
@@ -377,7 +400,7 @@ def compute_and_write_additional_sstat_fields(
         do_dssum_on_3comp_vector(dQ2_dxi, msh_conn, msh)
         do_dssum_on_3comp_vector(dQ3_dxi, msh_conn, msh)
 
-    this_file_name = which_dir + "/dUSdx" + this_ext
+    this_file_name = which_dir + "/dUiSdxj" + this_ext
     write_file_9c(
         comm,
         msh,
@@ -391,7 +414,7 @@ def compute_and_write_additional_sstat_fields(
     stat_fields.clear()
 
     ###########################################################################################
-    # UiSS first derivative
+    # UiSS divergence
     ###########################################################################################
 
     stat_fields.add_field(
@@ -407,7 +430,7 @@ def compute_and_write_additional_sstat_fields(
         field_name="VSS",
         file_type="fld",
         file_name=full_fname_stat,
-        file_key=file_keys_UiS[1],
+        file_key=file_keys_UiSS[1],
         dtype=np.single,
     )
     stat_fields.add_field(
@@ -434,50 +457,179 @@ def compute_and_write_additional_sstat_fields(
         do_dssum_on_3comp_vector(dQ2_dxi, msh_conn, msh)
         do_dssum_on_3comp_vector(dQ3_dxi, msh_conn, msh)
 
-    this_file_name = which_dir + "/dUSSdx" + this_ext
-    write_file_9c(
+    this_file_name = which_dir + "/dUjSSdxj" + this_ext
+    write_file_3c(
         comm,
         msh,
-        dQ1_dxi,
-        dQ2_dxi,
-        dQ3_dxi,
+        dQ1_dxi.c1 + dQ2_dxi.c2 + dQ3_dxi.c3,
         this_file_name,
         if_write_mesh=if_write_mesh,
     )
 
     stat_fields.clear()
-    del dQ2_dxi, dQ3_dxi
 
     ###########################################################################################
-    # UiUjS first derivative
+    # UiUjS gradients: dUiUjS/dxj
     ###########################################################################################
 
[…]
