+++ pysemtools/postprocessing/statistics/passive_scalar_budgets_interpolatedPoints_notMPI.py
@@ -4,6 +4,9 @@
 #####################################################################################
 # function to read and store the interpolated fields as structured and averaged fields
 #####################################################################################
+import time
+
+
 def read_interpolated_scalar_stat_hdf5_fields(
     path_to_files,
     Reynolds_number,
@@ -83,9 +86,11 @@ def read_interpolated_scalar_stat_hdf5_fields(
     dS2dxj_vec = np.column_stack([load_field(i) for i in range(48, 51)])
     d2S2dxj2_vec = np.column_stack([load_field(i) for i in range(51, 54)])
     dUiSdxj_vec = np.column_stack([load_field(i) for i in range(54, 63)])
-    dUiS2dxj_vec = np.column_stack([load_field(i) for i in range(63, 72)])
-    dUiUjSdxk_vec = np.column_stack([load_field(i) for i in range(72, 90)])
-    dPSdxj_vec = np.column_stack([load_field(i) for i in range(90, 93)])
+    dUjS2dxj_vec = np.column_stack([load_field(63)])
+    dUiUjSdxj_vec = np.column_stack([load_field(i) for i in range(64, 67)])
+    dPSdxj_vec = np.column_stack([load_field(i) for i in range(67, 70)])
+    dUidSdxjdxj_vec = np.column_stack([load_field(i) for i in range(70, 73)])
+    dSdUidxjdxj_vec = np.column_stack([load_field(i) for i in range(73, 76)])
 
     print("Finished reading all fields.")
     print(f"Done in {time.time() - start_time:.2f} seconds.")
@@ -246,11 +251,19 @@ def calculate_scalar_budgets_in_Cartesian(
 
         print(f"Done in {time.time() - start_time:.2f} seconds.")
 
-        # %% Mean velocity derivatives
-        print("--------------loading mean velocity derivatives...")
+        # %% Mean velocity gradients
+        print("--------------loading mean velocity gradients...")
         start_time = time.time()
 
         dUidXj = np.array(input_fluid_file["dUidxj_struct"])
+
+        print(f"Done in {time.time() - start_time:.2f} seconds.")
+
+        # %% Mean velocity second derivatives
+        print("--------------loading mean velocity second derivatives...")
+        start_time = time.time()
+
+        d2UidXj2 = np.array(input_fluid_file["d2Uidx2_struct"])
 
         print(f"Done in {time.time() - start_time:.2f} seconds.")
 
@@ -259,6 +272,14 @@ def calculate_scalar_budgets_in_Cartesian(
         start_time = time.time()
 
         P = np.array(input_fluid_file["P_struct"])
+
+        print(f"Done in {time.time() - start_time:.2f} seconds.")
+
+        # %% Pressure gradients
+        print("--------------loading pressure gradients...")
+        start_time = time.time()
+
+        dPdXj = np.array(input_fluid_file["dPdxj_struct"])
 
         print(f"Done in {time.time() - start_time:.2f} seconds.")
 
@@ -379,6 +400,8 @@ def calculate_scalar_budgets_in_Cartesian(
 
         S_convection = np.sum(Ui * dSdXj, axis=-1)
 
+        S_convection = -S_convection  # Put on RHS
+
[…]
