+++ pysemtools/monitoring/__init__.py
@@ -1,6 +1,5 @@
 """ Monitoring tools for pySEMTools """
 
 from .logger import Logger
-from .memory_monitor import MemoryMonitor
 
-__all__ = ["Logger", "MemoryMonitor"]
+__all__ = ["Logger"]
