+++ pysemtools/postprocessing/statistics/passive_scalar_budgets.py
@@ -112,6 +112,24 @@ def write_file_3c(comm, msh, dU_dxi, fname_gradU, if_write_mesh):
 
     gradU.clear()
 
+###########################################################################################
+###########################################################################################
+
+
+# %% generic function to write the sum of a 3-component (vector) field (as a scalar)
+def write_file_sum3c(comm, msh, dU_dxi, fname_gradU, if_write_mesh):
+    from pysemtools.datatypes.field import FieldRegistry
+    from pysemtools.io.ppymech.neksuite import pynekwrite
+    import numpy as np
+
+    gradU = FieldRegistry(comm)
+
+    gradU.add_field(comm, field_name="sum", field=dU_dxi.c1 + dU_dxi.c2 + dU_dxi.c3, dtype=np.single)
+
+    pynekwrite(fname_gradU, comm, msh=msh, fld=gradU, wdsz=4, write_mesh=if_write_mesh)
+
+    gradU.clear()
+
 
 ###########################################################################################
 ###########################################################################################
@@ -140,10 +158,11 @@ def return_list_of_vars_from_filename(fname):
         scal_fields = 37
 
     field_names = []
+
     for i in range(vel_fields):
         tmp = ["vel_" + str(i)]
         field_names = field_names + tmp
-
+    
     if pres_fields == 1:
         field_names = field_names + ["pres"]
 
@@ -175,12 +194,73 @@ def do_dssum_on_3comp_vector(dU_dxi, msh_conn, msh):
 ###########################################################################################
 ###########################################################################################
 
+#%% convert 2D statistics to 3D
+def convert_scalar_2Dstats_to_3D(stats2D_filename,stats3D_filename,datatype='single'):
+    from mpi4py import MPI 
+    import numpy as np
+    import warnings
+    from pysemtools.io.ppymech.neksuite import pynekread,pynekwrite
+    from pysemtools.datatypes.msh import Mesh
+    from pysemtools.datatypes.field import FieldRegistry
+    from pysemtools.datatypes.utils import extrude_2d_sem_mesh
+    from pysemtools.interpolation.interpolator import get_bbox_from_coordinates
+    from pysemtools.interpolation.point_interpolator.single_point_helper_functions import GLL_pwts
+
+    # Get mpi info
+    comm = MPI.COMM_WORLD
+
+    # initialize fields
+    msh = Mesh(comm, create_connectivity=False)
+    fld = FieldRegistry(comm)
+
+    if datatype=='single':
+        data_type = np.single
+        wdsz = 4
+    else:
+        data_type = np.double
+        wdsz = 8
+
+    # read mesh and fields
+    pynekread(stats2D_filename, comm, data_dtype=data_type, msh=msh, fld=fld)
+
+    # get the 
+    field_names = return_list_of_vars_from_filename(stats2D_filename)
+    print('fields in the given file: ', field_names )
+
+    # extruding mesh and fields
+    ## Find how much to extrude - Extrude the size of the smallest element
+    bbox = get_bbox_from_coordinates(msh.x, msh.y, msh.z)
+    bbox_dist = np.zeros((bbox.shape[0], 3))
+    bbox_dist[:, 0] = bbox[:, 1] - bbox[:, 0]
+    bbox_dist[:, 1] = bbox[:, 3] - bbox[:, 2]
+    bbox_dist[:, 2] = bbox[:, 5] - bbox[:, 4]
+    local_bbox_min_dist = np.min(
[…]
