+++ pysemtools/postprocessing/statistics/passive_scalar_budgets_interpolatedPoints_notMPI.py
@@ -55,18 +55,18 @@ def read_interpolated_scalar_stat_hdf5_fields(
     print("reading the 42 statistics fields...")
     start_time = time.time()
 
-    S_vec = load_field(0)
+    S_vec = load_field(4)
     UiS_vec = np.column_stack([load_field(i) for i in range(1, 4)])
-    S2_vec = load_field(4)
-    S3_vec = load_field(5)
-    S4_vec = load_field(6)
-    UiS2_vec = np.column_stack([load_field(i) for i in range(7, 10)])
-    UiUjS_vec = np.column_stack([load_field(i) for i in range(10, 16)])
-    PS_vec = load_field(16)
-    PdSdxi_vec = np.column_stack([load_field(i) for i in range(17, 20)])
-    UidSdxj_vec = np.column_stack([load_field(i) for i in range(20, 29)])
-    SdUidxj_vec = np.column_stack([load_field(i) for i in range(29, 38)])
-    SDiss_vec = np.column_stack([load_field(i) for i in range(38, 42)])
+    S2_vec = load_field(5)
+    S3_vec = load_field(6)
+    S4_vec = load_field(7)
+    UiS2_vec = np.column_stack([load_field(i) for i in range(8, 11)])
+    UiUjS_vec = np.column_stack([load_field(i) for i in range(11, 17)])
+    PS_vec = load_field(17)
+    PdSdxi_vec = np.column_stack([load_field(i) for i in range(18, 21)])
+    UidSdxj_vec = np.column_stack([load_field(i) for i in range(21, 30)])
+    SdUidxj_vec = np.column_stack([load_field(i) for i in range(30, 39)])
+    SDiss_vec = np.column_stack([load_field(i) for i in range(39, 43)])
 
     print(f"Done in {time.time() - start_time:.2f} seconds.")
 
@@ -81,16 +81,16 @@ def read_interpolated_scalar_stat_hdf5_fields(
     print("reading the additional fields...")
     start_time = time.time()
 
-    dSdxj_vec = np.column_stack([load_field(i) for i in range(42, 45)])
-    d2Sdxj2_vec = np.column_stack([load_field(i) for i in range(45, 48)])
-    dS2dxj_vec = np.column_stack([load_field(i) for i in range(48, 51)])
-    d2S2dxj2_vec = np.column_stack([load_field(i) for i in range(51, 54)])
-    dUiSdxj_vec = np.column_stack([load_field(i) for i in range(54, 63)])
-    dUjS2dxj_vec = np.column_stack([load_field(63)])
-    dUiUjSdxj_vec = np.column_stack([load_field(i) for i in range(64, 67)])
-    dPSdxj_vec = np.column_stack([load_field(i) for i in range(67, 70)])
-    dUidSdxjdxj_vec = np.column_stack([load_field(i) for i in range(70, 73)])
-    dSdUidxjdxj_vec = np.column_stack([load_field(i) for i in range(73, 76)])
+    dSdxj_vec = np.column_stack([load_field(i) for i in range(43, 46)])
+    d2Sdxj2_vec = np.column_stack([load_field(i) for i in range(46, 49)])
+    dS2dxj_vec = np.column_stack([load_field(i) for i in range(49, 52)])
+    d2S2dxj2_vec = np.column_stack([load_field(i) for i in range(52, 55)])
+    dUiSdxj_vec = np.column_stack([load_field(i) for i in range(55, 64)])
+    dUjS2dxj_vec = np.column_stack([load_field(64)])
+    dUiUjSdxj_vec = np.column_stack([load_field(i) for i in range(65, 68)])
[…]
