### 2026-02-23 ffdd374
Added write message
