### 2026-02-24 94c9c99
fix kwargs for find_points_tol
