### 2026-02-24 38e292b
Rename project to extremeflow-pysemtools and update version
