### 2026-02-24 f405ff2
Revise installation and dependencies in README
Updated installation instructions for mpi4py and PyTorch, clarified PyPI installation options, and modified dependencies section.
