### 2026-02-24 dbe9471
Fix typos and enhance README clarity
Correct typos and improve clarity in README.
