### 2026-02-25 ExtremeFLOW/pySEMTools#238 @adperezm
Okay!
The contributions of the authors of the paper are as follows:

**Adalberto Perez** is the main contributor, in charge of implementing most of the functionalities.

**Siavash Toosi** Implemented some of the post-processing methods we use in the examples (particularly those related to statistics). He also made part of the testing related to correctness and performance of most of the objects and procedures defined in the package.

**Tim Felle Olsen** implemented one of the readers available to consume data. Furthermore he helped in designing the testing-suite and in the overall design of the package.

**Philipp Schlatter** and **Stefano Markidis** provided direction on the type of tools and methods that were implemented such that they would truly provide needed functionalities in the larger context of simulations using SEM and high-performance computing. Furthermore they helped in assessing the theoretical basis of the approaches we followed and the correctness of the implementations.
