### 2026-02-26 22ed2b8
Rework overwrite prompt and fix index key handling
