### 2026-02-26 1610e76
remove redundant file exists check
