### 2026-03-01 ExtremeFLOW/pySEMTools#235 @finsberg
Hi @adperezm ! This is really nice work! I might be a good idea to include some more documentation on the landing page for the pypi package. Right now there is nothing there: https://pypi.org/project/extremeflow-pysemtools/

A common thing to do is to add the [readme field](https://peps.python.org/pep-0621/#readme) to your `pyproject.toml`
