+++ pyproject.toml
@@ -5,6 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "extremeflow-pysemtools"
 version = "1.2.0"
+readme = "README.md"
 requires-python = ">= 3.10"
 dependencies = [
     "numpy",
