### 2026-03-02 78c2488
fix d(SdW/dxj)/dxj processing
