+++ pysemtools/io/wrappers.py
@@ -1,5 +1,6 @@
 """Wrappers to ease IO"""
 
+import sys
 from typing import Union
 import os
 import h5py
@@ -10,6 +11,20 @@ from .ppymech.neksuite import pynekread, pynekwrite
 import numpy as np
 from mpi4py import MPI
 from ..monitoring.logger import Logger
+
+def ijk_to_id(i, j, k, ny, nz):
+    return i*(ny*nz) + j*nz + k
+
+def ijk_to_id_mpi(i, j, k, ny, nz, parallel_axis, offset):
+    if parallel_axis == 0:
+        i = i + offset
+        return i*(ny*nz) + j*nz + k
+    else:
+        raise NotImplementedError("Parallel axis other than 0 is not implemented for the ijk_to_id_mpi function")
+
+def write_headers(root):
+    root.attrs["Version"] = (2, 3)
+    root.attrs["Type"] = "UnstructuredGrid"
 
 def partition_read_data(comm, fname: str = None, distributed_axis: int = 0):
     """
@@ -202,7 +217,77 @@ def read_data(comm, fname: str, keys: list[str], parallel_io: bool = False, dtyp
                 # Read the field
                 fld.add_field(comm, field_name=key, file_type="fld", file_name=fname, file_key=key, dtype=dtype)
                 data[key] = np.copy(fld.registry[key])
-                fld.clear()         
+                fld.clear() 
+
+    elif extension == 'vtkhdf':        
+        if parallel_io:
+            if distributed_axis != 0:
+                raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
+            
+            with h5py.File(fname, 'r', driver='mpio', comm=comm) as f:
+                data = {}
+                for key in keys:
+
+                    global_array_shape = tuple(f["VTKHDF"]["PointData"][key].attrs["real_global_shape"])
+                    # Determine how many axis zero elements to get locally
+                    # This corresponds to a linearly load balanced partitioning
+                    i_rank = comm.Get_rank()
+                    m = global_array_shape[distributed_axis]
+                    pe_rank = i_rank
+                    pe_size = comm.Get_size()
+                    ip = np.floor(
+                        (
+                            np.double(m)
+                            + np.double(pe_size)
+                            - np.double(pe_rank)
+                            - np.double(1)
+                        )
+                        / np.double(pe_size)
+                    )
+                    local_axis_0_shape = int(ip)
+                    #determine the offset
+                    offset = comm.scan(local_axis_0_shape) - local_axis_0_shape
+
+                    # The offset is now in number of rows (or first axis). But it needs to be into the flattened array
+                    if distributed_axis == 0:
+                        offset = offset * (global_array_shape[1] * global_array_shape[2])
+                    else:
+                        raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
+
+                    # Determine the local array shape
+                    temp = list(global_array_shape)
+                    temp[distributed_axis] = local_axis_0_shape
+                    local_array_shape = tuple(temp)
+                    
+                    # Get the slice of the data that should be read
+                    slices = [slice(None)]
+                    if distributed_axis == 0:
+                        slices[distributed_axis] = slice(offset, offset + local_array_shape[distributed_axis]*(global_array_shape[1] * global_array_shape[2]))
+                    else:
+                        raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
+
[…]
