### 2026-03-02 d8a7ef8
Added draft hdf5 object to manage that type of io
