+++ pysemtools/io/hdf/__init__.py
@@ -1,0 +1,1 @@
+""" Classes to read or write data in the hdf format"""
