+++ pysemtools/io/hdf/hdf5.py
@@ -1,0 +1,168 @@
+""" Module that defines the hdf5 object to be used in pysemtools"""
+
+import os
+import h5py
+import numpy as np
+from mpi4py import MPI
+
+class HDF5File:
+    """
+    Class to write and read hdf5 files in parallel using h5py.
+    """
+
+    def __init__(self, comm : MPI.Comm, fname: str, mode: str, parallel: bool):
+        """ Initialize the hdf5 file object
+        
+        Open an hdf5 file based on inputs.
+
+        Parameters
+        ----------
+        comm : MPI.Comm
+            MPI communicator to use for parallel I/O. If parallel is False, this can be set
+            to None.
+        fname : str
+            Name of the hdf5 file to read or write.
+        mode : str
+            Mode to open the file. Should be "read" for reading or "write" for
+            writing.
+        parallel : bool
+            Whether to use parallel I/O or not. 
+        """
+        
+        self.comm = comm
+        self.fname = fname
+        if mode not in ["r", "w"]:
+            raise ValueError("Mode should be 'read' or 'write'")
+        else:
+            self.mode = mode
+        self.parallel = parallel
+
+        if self.parallel:
+            self.file = h5py.File(self.fname, self.mode, driver='mpio', comm=self.comm)
+        else:
+            self.file = h5py.File(self.fname, self.mode)
+
+        # Some temporary variables to store
+        self.global_shape = None
+        self.local_shape = None
+        self.offset = None
+        self.count = None
+        self.slices = None
+        self.local_alloc_shape = None
+
+    def close(self, clean: bool = False):
+        """ Close the hdf5 file object """
+        self.file.close()
+
+        if clean:
+            self.global_shape = None
+            self.local_shape = None
+            self.offset = None
+            self.count = None
+            self.slices = None        
+            self.local_alloc_shape = None
+
+
+    def read_dataset(self, dataset_name: str, dtype : np.dtype = np.double, distributed_axis: int = None, slices: list = None):
+        """ Read a dataset from the hdf5 file object 
+        """
+
+        if self.mode != "r":
+            raise ValueError("File is not opened in read mode")
+
+        if self.parallel and distributed_axis is None:
+            raise ValueError("Distributed axis must be specified for parallel reading")
+                
+        if slices is not None and len(slices) != self.file[dataset_name].ndim:
+            raise ValueError("Number of slices must match the number of dimensions of the dataset")
+
+        # ===========
+        # Serial read 
+        # ===========
+        if not self.parallel:
+            if slices is None:
+                data = self.file[dataset_name][:]
+            else:
+                data = self.file[dataset_name][tuple(slices)]
+        
+        # =============
+        # Parallel read 
+        # =============
+        else:
+            if slices is None:
+                self.set_read_slices(global_shape=self.file[dataset_name].shape, distributed_axis=distributed_axis, explicit_strides=False)
+            else:
+                # If the slices are explicitly provided, determine the local shape to allocate
+                self.global_shape = self.file[dataset_name].shape
+                
+                # Local shape from slices
+                local_array_shape = []
+                for dim, slc in zip(self.global_shape, slices):
+                    if isinstance(slc, slice):
+                        start = 0 if slc.start is None else slc.start
+                        stop = dim if slc.stop is None else slc.stop
+                        local_array_shape.append(stop - start)
+
+                # Set the attributes
+                self.local_shape = tuple(local_array_shape)
+                self.local_alloc_shape = self.local_shape
+                self.slices = tuple(slices)
+
+            local_data = np.empty(self.local_alloc_shape, dtype=dtype)
+            local_data[:] = self.file[dataset_name][self.slices].reshape(self.local_shape)
+
+        return local_data 
[…]
