### 2026-03-02 388af33
Added write functions to hdf class
