+++ pysemtools/io/hdf/hdf5.py
@@ -28,8 +28,30 @@ class HDF5File:
         parallel : bool
             Whether to use parallel I/O or not. 
         """
-        
+
+        # Assign the communicator and assign empty attributes        
         self.comm = comm
+        
+        # Set attributes that are assigned when opening a file
+        self.mode = None
+        self.parallel = None
+        self.file = None
+        self.fname = None
+
+        # Some temporary variables to store
+        self.global_shape = None
+        self.local_shape = None
+        self.offset = None
+        self.count = None
+        self.slices = None
+        self.local_alloc_shape = None
+
+        # Open a file
+        self.open(fname, mode, parallel)
+
+    def open(self, fname: str, mode: str, parallel: bool):
+        """ Open an hdf5 file based on inputs. This can be used to open a new file after closing the previous one."""
+        
         self.fname = fname
         if mode not in ["r", "w"]:
             raise ValueError("Mode should be 'read' or 'write'")
@@ -41,14 +63,6 @@ class HDF5File:
             self.file = h5py.File(self.fname, self.mode, driver='mpio', comm=self.comm)
         else:
             self.file = h5py.File(self.fname, self.mode)
-
-        # Some temporary variables to store
-        self.global_shape = None
-        self.local_shape = None
-        self.offset = None
-        self.count = None
-        self.slices = None
-        self.local_alloc_shape = None
 
     def close(self, clean: bool = False):
         """ Close the hdf5 file object """
@@ -89,33 +103,21 @@ class HDF5File:
         # Parallel read 
         # =============
         else:
+            # Set slices
             if slices is None:
-                self.set_read_slices(global_shape=self.file[dataset_name].shape, distributed_axis=distributed_axis, explicit_strides=False)
+                self.set_read_slices_linear_lb(global_shape=self.file[dataset_name].shape, distributed_axis=distributed_axis, explicit_strides=False)
             else:
-                # If the slices are explicitly provided, determine the local shape to allocate
-                self.global_shape = self.file[dataset_name].shape
-                
-                # Local shape from slices
-                local_array_shape = []
-                for dim, slc in zip(self.global_shape, slices):
-                    if isinstance(slc, slice):
-                        start = 0 if slc.start is None else slc.start
-                        stop = dim if slc.stop is None else slc.stop
-                        local_array_shape.append(stop - start)
+                self.set_read_slices_external(global_shape=self.file[dataset_name].shape, slices=slices)
+            
+            local_data = self.read_slices(dataset_name, dtype=dtype)
 
-                # Set the attributes
-                self.local_shape = tuple(local_array_shape)
-                self.local_alloc_shape = self.local_shape
-                self.slices = tuple(slices)
-
-            local_data = np.empty(self.local_alloc_shape, dtype=dtype)
-            local_data[:] = self.file[dataset_name][self.slices].reshape(self.local_shape)
-
-        return local_data 
+        return local_data
     
-    def set_read_slices(self, global_shape: tuple, distributed_axis: int, explicit_strides: bool = False):
-        """Set the slices that should be read from the file."""
+    def set_read_slices_linear_lb(self, global_shape: tuple, distributed_axis: int, explicit_strides: bool = False):
+        """Set the slices that should be read from the file.
 
+        Data is distributed in a linear load balanced way.
+        """
         # Perform a load balanced distribution
         i_rank = self.comm.Get_rank()
         m = global_shape[distributed_axis]
@@ -166,3 +168,104 @@ class HDF5File:
         self.slices = tuple(slices)
         self.local_shape = local_shape
         self.local_alloc_shape = local_alloc_shape
+    
+    def set_read_slices_external(self, global_shape: tuple, slices: list):
[…]
