+++ pysemtools/io/wrappers.py
@@ -11,6 +11,7 @@ from .ppymech.neksuite import pynekread, pynekwrite
 import numpy as np
 from mpi4py import MPI
 from ..monitoring.logger import Logger
+from .hdf.hdf5 import HDF5File
 
 def ijk_to_id(i, j, k, ny, nz):
     return i*(ny*nz) + j*nz + k
@@ -118,81 +119,13 @@ def read_data(comm, fname: str, keys: list[str], parallel_io: bool = False, dtyp
 
     # Read the data
     if extension == 'hdf5':
-        if parallel_io:
-            #raise NotImplementedError("Parallel IO is not implemented for hdf5 files")
-            with h5py.File(fname, 'r', driver='mpio', comm=comm) as f:
-                data = {}
-                for key in keys:
-
-                    # If the slices are not provided, determine the local data to read
-                    if slices is None:
-
-                        # Get the global array shape and sizes
-                        global_array_shape = f[key].shape
-                        # Determine how many axis zero elements to get locally
-                        # This corresponds to a linearly load balanced partitioning
-                        i_rank = comm.Get_rank()
-                        m = global_array_shape[distributed_axis]
-                        pe_rank = i_rank
-                        pe_size = comm.Get_size()
-                        ip = np.floor(
-                            (
-                                np.double(m)
-                                + np.double(pe_size)
-                                - np.double(pe_rank)
-                                - np.double(1)
-                            )
-                            / np.double(pe_size)
-                        )
-                        local_axis_0_shape = int(ip)
-                        #determine the offset
-                        offset = comm.scan(local_axis_0_shape) - local_axis_0_shape
-
-                        # Determine the local array shape
-                        temp = list(global_array_shape)
-                        temp[distributed_axis] = local_axis_0_shape
-                        local_array_shape = tuple(temp)
-                        
-                        # Get the slice of the data that should be read
-                        slices = [slice(None) for i in range(len(global_array_shape))]
-                        slices[distributed_axis] = slice(offset, offset + local_array_shape[distributed_axis])                    
-
-                    # Otherwise, if the slices are provided, use them to read the data.
-                    else:
-
-                        # Get the global array shape and sizes
-                        global_array_shape = f[key].shape
-
-                        local_array_shape = []
-                        for i in range(len(global_array_shape)):
-                            if isinstance(slices[i], slice) and slices[i].start is not None and slices[i].stop is not None:
-                                local_array_shape.append(slices[i].stop - slices[i].start)
-                            elif isinstance(slices[i], slice) and slices[i].start is None and slices[i].stop is None:
-                                local_array_shape.append(global_array_shape[i])
-
-                        local_array_shape = tuple(local_array_shape)
-                    
-                    # Allocate the local data array
-                    local_data = np.empty(local_array_shape, dtype=dtype)
-
-                    # Read the data
-                    local_data[:] = f[key][tuple(slices)]
-
-                    # Store the data
-                    data[key] = local_data
-
-        else:
-            with h5py.File(fname, 'r') as f:
-                data = {}
-                for key in keys:
-                    if slices is None:
-                        data[key] = f[key][:]
-                    else:
-                        if not len(slices) == f[key].ndim:
-                            raise ValueError("The number of slices provided does not match the number of dimensions in the data.")
-                        else:
-                            data[key] = f[key][tuple(slices)]
-    
+        
+        file = HDF5File(comm, fname, "r", parallel_io)
+        data = {}
+        for key in keys:
[…]
