+++ pysemtools/io/hdf/vtkhdf.py
@@ -1,0 +1,231 @@
+""" Module that defines the hdf5 object to be used in pysemtools"""
+
+import os
+import h5py
+import numpy as np
+from mpi4py import MPI
+from ...monitoring.logger import Logger
+from .hdf5 import HDF5File
+
+class VTKHDF(HDF5File):
+    """
+    Class to write and read vtkhdf files in parallel using h5py.
+    """
+
+    def __init__(self, comm : MPI.Comm, fname: str, mode: str, parallel: bool):
+        """ Initialize the hdf5 file object
+        
+        Open an hdf5 file based on inputs.
+
+        Parameters
+        ----------
+        comm : MPI.Comm
+            MPI communicator to use for parallel I/O. If parallel is False, this can be set
+            to None.
+        fname : str
+            Name of the hdf5 file to read or write.
+        mode : str
+            Mode to open the file. Should be "read" for reading or "write" for
+            writing.
+        parallel : bool
+            Whether to use parallel I/O or not. 
+        mesh : dict
+            Dictionary containing the mesh information. Should contain the keys "x", "y", and "z" for the coordinates of the mesh points.
+        """
+        super().__init__(comm, fname, mode, parallel)
+
+        self.shape = None
+        
+        # Write the headers if we are writing
+        self.set_active_group("/VTKHDF")
+        if self.mode == "w":
+            write_headers(self.active_group)
+
+    def write_mesh_data(self, mesh: dict, distributed_axis: int = 0):
+        """ Write the mesh data to the hdf5 file
+        
+        Parameters
+        ----------
+        mesh : dict
+            Dictionary containing the mesh information. Should contain the keys "x", "y", and "z" for the coordinates of the mesh points.
+        """
+
+        if distributed_axis != 0:
+            raise NotImplementedError("Distributed axis other than 0 is not implemented for the write_mesh_data function")
+
+        # =================
+        # Serial vtk set up
+        # =================
+        if not self.parallel:
+            vtk_data = set_up_hex_vtk_mesh_serial(mesh["x"], mesh["y"], mesh["z"])
+        else:
+            vtk_data = set_up_hex_vtk_mesh_parallel(self.comm, mesh["x"], mesh["y"], mesh["z"], distributed_axis)
+
+        # Write the mesh metadata
+        self.set_active_group("/VTKHDF")
+        self.active_group.create_dataset("NumberOfPoints", data=vtk_data.pop("NumberOfPoints"), dtype="i8")
+        self.active_group.create_dataset("NumberOfCells", data=vtk_data.pop("NumberOfCells"), dtype="i8")
+        nconectivity_ids = vtk_data["NumberOfConnectivityIds"][0]
+        self.active_group.create_dataset("NumberOfConnectivityIds", data=vtk_data.pop("NumberOfConnectivityIds"), dtype="i8")
+
+        # Save the shape of the mesh for later use
+        self.shape = vtk_data.pop("shape")
+
+        # Write the mesh data
+        self.write_dataset("Points", vtk_data.pop("Points"), distributed_axis=distributed_axis)
+        self.write_dataset("Connectivity", vtk_data.pop("Connectivity"), distributed_axis=distributed_axis)
[…]
