### 2026-03-03 8ac796e
Finalized wrapper integration
