+++ pysemtools/io/wrappers.py
@@ -12,6 +12,7 @@ import numpy as np
 from mpi4py import MPI
 from ..monitoring.logger import Logger
 from .hdf.hdf5 import HDF5File
+from .hdf.vtkhdf import VTKHDFFile
 
 def ijk_to_id(i, j, k, ny, nz):
     return i*(ny*nz) + j*nz + k
@@ -153,73 +154,13 @@ def read_data(comm, fname: str, keys: list[str], parallel_io: bool = False, dtyp
                 fld.clear() 
 
     elif extension == 'vtkhdf':        
-        if parallel_io:
-            if distributed_axis != 0:
-                raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
-            
-            with h5py.File(fname, 'r', driver='mpio', comm=comm) as f:
-                data = {}
-                for key in keys:
-
-                    global_array_shape = tuple(f["VTKHDF"]["PointData"][key].attrs["global_shape"])
-                    # Determine how many axis zero elements to get locally
-                    # This corresponds to a linearly load balanced partitioning
-                    i_rank = comm.Get_rank()
-                    m = global_array_shape[distributed_axis]
-                    pe_rank = i_rank
-                    pe_size = comm.Get_size()
-                    ip = np.floor(
-                        (
-                            np.double(m)
-                            + np.double(pe_size)
-                            - np.double(pe_rank)
-                            - np.double(1)
-                        )
-                        / np.double(pe_size)
-                    )
-                    local_axis_0_shape = int(ip)
-                    #determine the offset
-                    offset = comm.scan(local_axis_0_shape) - local_axis_0_shape
-
-                    # The offset is now in number of rows (or first axis). But it needs to be into the flattened array
-                    if distributed_axis == 0:
-                        offset = offset * (global_array_shape[1] * global_array_shape[2])
-                    else:
-                        raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
-
-                    # Determine the local array shape
-                    temp = list(global_array_shape)
-                    temp[distributed_axis] = local_axis_0_shape
-                    local_array_shape = tuple(temp)
-                    
-                    # Get the slice of the data that should be read
-                    slices = [slice(None)]
-                    if distributed_axis == 0:
-                        slices[distributed_axis] = slice(offset, offset + local_array_shape[distributed_axis]*(global_array_shape[1] * global_array_shape[2]))
-                    else:
-                        raise NotImplementedError("Parallel axis other than 0 is not implemented for reading vtkhdf files in parallel")
-
-                    
-                    # Allocate the local data array
-                    local_data = np.empty(local_array_shape, dtype=dtype)
-                    local_data = local_data.reshape(-1)
-
-                    # Read the data
-                    local_data[:] = f["VTKHDF"]["PointData"][key][tuple(slices)] 
-                    
-                    # Store the data
-                    data[key] = local_data.reshape(local_array_shape)
-
-        else:
-            with h5py.File(fname, 'r') as f:
-                data = {}
-                for key in keys:
-                    temp_data = f["VTKHDF"]["PointData"][key][:]
-                    if "global_shape" in f["VTKHDF"]["PointData"][key].attrs:
-                        original_shape = f["VTKHDF"]["PointData"][key].attrs["global_shape"]
[…]
