### 2026-03-03 869d2c4
Change bibitem from arxiv version to published version
