+++ paper.bib
@@ -29,10 +29,17 @@ abstract = {A spectral element method that combines the generality of the finite
 }
 
 @article{fischer2021nekrs,
-  title={NekRS, a GPU-Accelerated Spectral Element Navier-Stokes Solver},
-  author={Fischer, Paul and Kerkemeier, Stefan and Min, Misun and Lan, Yu-Hsiang and Phillips, Malachi and Rathnayake, Thilina and Merzari, Elia and Tomboulides, Ananias and Karakus, Ali and Chalmers, Noel and others},
-  journal={arXiv preprint arXiv:2104.05829},
-  year={2021}
+title = {NekRS, a GPU-accelerated spectral element Navier-Stokes solver},
+journal = {Parallel Computing},
+volume = {114},
+pages = {102982},
+year = {2022},
+issn = {0167-8191},
+doi = {https://doi.org/10.1016/j.parco.2022.102982},
+url = {https://www.sciencedirect.com/science/article/pii/S0167819122000710},
+author = {Paul Fischer and Stefan Kerkemeier and Misun Min and Yu-Hsiang Lan and Malachi Phillips and Thilina Rathnayake and Elia Merzari and Ananias Tomboulides and Ali Karakus and Noel Chalmers and Tim Warburton},
+keywords = {NekRS, Nek5000, LibParanumal, OCCA, GPU, Scalability, Performance, Spectral element method, Incompressible Navier–Stokes, Exascale applications},
+abstract = {The development of NekRS, a GPU-oriented thermal-fluids simulation code based on the spectral element method (SEM) is described. For performance portability, the code is based on the open concurrent compute abstraction and leverages scalable developments in the SEM code Nek5000 and in libParanumal, which is a library of high-performance kernels for high-order discretizations and PDE-based miniapps. Critical performance sections of the Navier–Stokes time advancement are addressed. Performance results on several platforms are presented, including scaling to 27,648 V100s on OLCF Summit, for calculations of up to 60B grid points (240B degrees-of-freedom).}
 }
 
 @article{jansson2024,
