### 2026-03-03 #8767 @adperezm
Hi @finsberg. It seems the feature/paper branch you updated is not in sync with the main branch. The pdf generated is not what you saw there, I think.

Let me check that all is as it should
