### 2026-03-03 3e0ea12
Added more docstrings
