+++ pysemtools/io/hdf/hdf5.py
@@ -24,7 +24,7 @@ class HDF5File:
         fname : str
             Name of the hdf5 file to read or write.
         mode : str
-            Mode to open the file. Should be "read" for reading or "write" for
+            Mode to open the file. Should be "r" for reading or "w" for
             writing.
         parallel : bool
             Whether to use parallel I/O or not. 
@@ -53,7 +53,16 @@ class HDF5File:
         self.open(fname, mode, parallel)
  
     def set_active_group(self, group_name: str):
-        """ Set the active group to read or write data from. This is useful to avoid having to specify the group every time a dataset is read or written."""
+        """ Set the active group to read or write data from. 
+        
+        This is useful to avoid having to specify the group every time a dataset is read or written.
+
+        Parameters
+        ----------
+        group_name : str
+            Name of the group to set as active. Can include the group path, e.g. "/group1/group2". 
+            If the group does not exist, it will be created if the file is opened in write mode, otherwise an error will be raised. 
+        """
         
         group_path = group_name.split("/")
         group_path = [name for name in group_path if name != ""] # Remove empty strings
@@ -71,12 +80,26 @@ class HDF5File:
         self.active_group = root
 
     def open(self, fname: str, mode: str, parallel: bool):
-        """ Open an hdf5 file based on inputs. This can be used to open a new file after closing the previous one."""
+        """ Open an hdf5 file based on inputs. 
+        
+        This can be used to open a new file after closing the previous one.
+        
+        Parameters
+        ----------
+        fname : str
+            Name of the hdf5 file to read or write.
+        mode : str
+            Mode to open the file. Should be "r" for reading or "w" for
+            writing.
+        parallel : bool
+            Whether to use parallel I/O or not. If True, the file will be opened using
+            the MPI-IO driver. If False, the file will be opened using the default driver.
+        """
 
         self.log.tic() 
         self.fname = fname
         if mode not in ["r", "w"]:
-            raise ValueError("Mode should be 'read' or 'write'")
+            raise ValueError("Mode should be 'r' or 'w'")
         else:
             self.mode = mode
  
@@ -96,7 +119,15 @@ class HDF5File:
         self.set_active_group("/")
 
     def close(self, clean: bool = False):
-        """ Close the hdf5 file object """
+        """ Close the hdf5 file object 
+        
+        Parameters
+        ----------
+        clean : bool
+            Whether to clean the attributes that are assigned when opening a file. This is useful if the
+            file object will be reused to open another file after closing the current one. Default is False.
+        
+        """
         self.file.close()
 
         if clean:
@@ -126,6 +157,12 @@ class HDF5File:
             Optional. default is False. Whether the data is stored as 1D in the file. This is useful if originally the data had a different shape
             but was flattened to 1d before writing. This will use the shape attribute stored in the file to do the partioning
             but will keep in mind that the data is stored as a 1d array to read properly.
+
+        Returns
+        -------
+        local_data : np.ndarray
+            Data read from the file. This will be a local array with the shape determined by the
+            global shape of the dataset and the parallel distribution. If slices are provided, the shape will be determined by the slices.
         """
 
         if self.mode != "r":
@@ -179,6 +216,18 @@ class HDF5File:
         """Set the slices that should be read from the file.
 
         Data is distributed in a linear load balanced way.
+
+        Parameters
+        ----------
+        global_shape : tuple
+            Shape of the global array to be read. This is required to determine the local shape and
+            the slices to read from the file.
+        distributed_axis : int
+            Axis along which the data is distributed in parallel. This is required to determine the local shape
+            and the slices to read from the file.
+        explicit_strides : bool
+            Whether to use explicit strides to read the data. This is useful if the data is stored
[…]
