### 2026-03-03 e6c0eed
Explicit mesh as inputs
