### 2026-03-03 2a9630e
Remove some helper functions in wrapper
