### 2026-03-03 2adef89
Added types to io init
