+++ pysemtools/io/hdf/__init__.py
@@ -1,1 +1,6 @@
 """ Classes to read or write data in the hdf format"""
+
+from.hdf5 import HDF5File
+from .vtkhdf import VTKHDFFile
+
+__all__ = ["HDF5File", "VTKHDFFile"]
