### 2026-03-04 #8767 @finsberg
@adperezm I think you need to update the doi from
```
doi = {doi.org/10.1016/j.parco.2022.102982},
```
to
```
doi = {10.1016/j.parco.2022.102982},
```
in https://github.com/ExtremeFLOW/pySEMTools/blob/6fcc5fd218537c5fe9a2cb18d03f4674e6bd9d26/paper.bib#L43C1-L43C45
