+++ pysemtools/io/hdf/hdf5.py
@@ -140,7 +140,7 @@ class HDF5File:
 
         self.log.toc(message=f"{self.fname} closed")
 
-    def read_dataset(self, dataset_name: str, dtype : np.dtype = np.double, distributed_axis: int = None, slices: list = None, as_1d_in_file: bool = False):
+    def read_dataset(self, dataset_name: str, dtype : np.dtype = np.double, distributed_axis: int = None, slices: list = None, as_array_list_in_file: bool = False):
         """ Read a dataset from the hdf5 file object
 
         Parameters
@@ -153,8 +153,8 @@ class HDF5File:
             Axis along which the data is distributed in parallel. This is required for parallel reading. Default is None.
         slices : list
             Optional. List of slices to read from the dataset. In case it is known
-        as_1d_in_file : bool
-            Optional. default is False. Whether the data is stored as 1D in the file. This is useful if originally the data had a different shape
+        as_array_list_in_file : bool
+            Optional. default is False. Whether the data is stored as an array list in the file. This is useful if originally the data had a different shape
             but was flattened to 1d before writing. This will use the shape attribute stored in the file to do the partioning
             but will keep in mind that the data is stored as a 1d array to read properly.
 
@@ -185,8 +185,10 @@ class HDF5File:
         # Query the shape
         if "shape" in self.active_group[dataset_name].attrs:
             global_shape = self.active_group[dataset_name].attrs["shape"]
+            shape_in_file = self.active_group[dataset_name].shape
         else:
             global_shape = self.active_group[dataset_name].shape
+            shape_in_file = self.active_group[dataset_name].shape
 
         # ===========
         # Serial read 
@@ -204,7 +206,7 @@ class HDF5File:
 
             # Set slices
             if slices is None:
-                self.set_read_slices_linear_lb(global_shape=global_shape, distributed_axis=distributed_axis, explicit_strides=as_1d_in_file)
+                self.set_read_slices_linear_lb(global_shape=global_shape, distributed_axis=distributed_axis, explicit_strides=as_array_list_in_file, shape_in_file=shape_in_file)
             else:
                 self.set_read_slices_external(global_shape=global_shape, slices=slices)
             
@@ -212,7 +214,7 @@ class HDF5File:
 
         return local_data
     
-    def set_read_slices_linear_lb(self, global_shape: tuple, distributed_axis: int, explicit_strides: bool = False):
+    def set_read_slices_linear_lb(self, global_shape: tuple, distributed_axis: int, explicit_strides: bool = False, shape_in_file: list = None):
         """Set the slices that should be read from the file.
 
         Data is distributed in a linear load balanced way.
@@ -250,9 +252,14 @@ class HDF5File:
 
         # Update the offset and count to traverse the non distributed axes if explicit strides are used
         if explicit_strides:
+            if shape_in_file is None:
+                raise ValueError("Shape in file must be provided if explicit strides are used")
+            # Determine which axes were merged
+            merged_axes = find_merged_axes(global_shape, shape_in_file) 
+            # This really only works relliably if distributed axis is 0 and is part of the merge
             stride = 1
             for i in range(len(global_shape)):
-                if i  != distributed_axis:
+                if i  != distributed_axis and merged_axes[i]:
                     stride = stride * global_shape[i]
             offset = offset * stride
             count = count * stride
@@ -264,9 +271,11 @@ class HDF5File:
 
         # build the slices to read
         if explicit_strides:
-            slices = [slice(None)]
+            slices = [slice(None)] * len(shape_in_file)
             slices[distributed_axis] = slice(offset, offset + count)
[…]
