### 2026-03-04 9a5a01a
Fixed typo and added exaple of new IO
