### 2026-03-04 101403b
Add the option to ignore metadata if needed
