### 2026-03-05 #8767 @finsberg
Hi @adperezm

Going through the final checklist here.

Could you please create a release of the software and post the version number here. Also archive this release to [Zenodo](https://zenodo.org/) (or another place that creates a DOI) and post the DOI here. It is important that the version you archive has the same title and authors as the JOSS paper (also the same LICESE as the repository).
