### 2026-03-05 #8767 @adperezm
Hi @finsberg, sounds good.

I have a question, just to confirm: when I submitted this paper we had v1.0.0, but now after some of these modifications we have bumped that version up. Will we be able to somehow change that? Or should I use the old ones?
