### 2026-03-05 #8767 @adperezm
> > Hi [@finsberg](https://github.com/finsberg), sounds good.
> > I have a question, just to confirm: when I submitted this paper we had v1.0.0, but now after some of these modifications we have bumped that version up. Will we be able to somehow change that? Or should I use the old ones?
>
> Yes, please make a new release with a new version. We can update the version in this submission to most most recent one (which has the changes from the review)

Excellent! I will get back to you as soon as I've finalized those.
