### 2026-03-05 86efdb5
Added dedicated type for vtk data
