+++ pysemtools/io/hdf/vtkhdf.py
@@ -6,6 +6,7 @@ import numpy as np
 from mpi4py import MPI
 from ...monitoring.logger import Logger
 from .hdf5 import HDF5File
+from ...datatypes import VTKMesh
 
 class VTKHDFFile(HDF5File):
     """
@@ -86,40 +87,28 @@ class VTKHDFFile(HDF5File):
         if distributed_axis != 0:
             raise NotImplementedError("Distributed axis other than 0 is not implemented for the write_mesh_data function")
 
-        # =================
-        # Serial vtk set up
-        # =================
-        if x.ndim == 3:
-            if not self.parallel:
-                vtk_data = set_up_hex_vtk_mesh_serial(x, y, z)
-            else:
-                vtk_data = set_up_hex_vtk_mesh_parallel(self.comm, x, y, z, distributed_axis)
-        elif x.ndim == 4:
-            if not self.parallel:
-                vtk_data = set_up_hex_vtk_mesh_sem_serial(x, y, z)
-            else:
-                vtk_data = set_up_hex_vtk_mesh_sem_parallel(self.comm, x, y, z)
-
-        else:
-            raise NotImplementedError("Only structured meshes with 3D/4D coordinates are currently supported for the write_mesh_data function")
+        # ========================
+        # Set up the vtk mesh data
+        # ========================
+        vtk = VTKMesh(self.comm, x, y, z, cell_type="hex", global_connectivity=self.parallel, distributed_axis=distributed_axis)
 
         # Write the mesh metadata
         self.set_active_group("/VTKHDF")
-        self.active_group.create_dataset("NumberOfPoints", data=vtk_data.pop("NumberOfPoints"), dtype="i8")
-        self.active_group.create_dataset("NumberOfCells", data=vtk_data.pop("NumberOfCells"), dtype="i8")
-        nconectivity_ids = vtk_data["NumberOfConnectivityIds"][0]
-        self.active_group.create_dataset("NumberOfConnectivityIds", data=vtk_data.pop("NumberOfConnectivityIds"), dtype="i8")
+        self.active_group.create_dataset("NumberOfPoints", data=vtk.data.pop("NumberOfPoints"), dtype="i8")
+        self.active_group.create_dataset("NumberOfCells", data=vtk.data.pop("NumberOfCells"), dtype="i8")
+        nconectivity_ids = vtk.data["NumberOfConnectivityIds"][0]
+        self.active_group.create_dataset("NumberOfConnectivityIds", data=vtk.data.pop("NumberOfConnectivityIds"), dtype="i8")
 
         # Save the shape of the mesh for later use
-        self.shape = vtk_data.pop("shape")
+        self.shape = vtk.data.pop("shape")
 
         # Write the mesh data
         mesh_shape = [axs for axs in self.shape] + [3]
-        self.write_dataset("Points", vtk_data.pop("Points"), distributed_axis=distributed_axis, shape_in_ram=mesh_shape)
-        self.write_dataset("Connectivity", vtk_data.pop("Connectivity"), distributed_axis=distributed_axis)
-        self.write_dataset("Types", vtk_data.pop("Types"), distributed_axis=distributed_axis)
[…]
