+++ pysemtools/datatypes/msh_vtk.py
@@ -1,0 +1,340 @@
+""" Module that contains vtk msh class, which contains relevant data on the domain"""
+
+__all__ = ['VTKMesh']
+
+import numpy as np
+from mpi4py import MPI
+    
+VTK_HEXAHEDRON = 12
+VTK_HEXAHEDRON_CELL_POINTS = 8
+
+class VTKMesh:
+    """ Class that contains the mesh data in a vtk friendly format
+    
+    Helper to build connectivity and offsets etc.
+
+    Parameters
+    ----------
+    comm : MPI.Comm
+        The MPI communicator
+    x : np.ndarray
+        The x coordinates of the mesh points
+    y : np.ndarray
+        The y coordinates of the mesh points
+    z : np.ndarray
+        The z coordinates of the mesh points
+    cell_type : str, optional
+        The type of cells in the mesh, by default "hex". Only "hex" is currently supported.
+    global_connectivity : bool, optional
+        Whether to use global connectivity or local connectivity in parallel. By default True, but if only one rank is used, it is set to False since the connectivity is already global in that case.
+    distributed_axis : int, optional
+        The axis along which the mesh is distributed, by default 0 (Only zero allowed now) 
+    """
+
+    def __init__(self, comm: MPI.Comm, x: np.ndarray, y: np.ndarray, z: np.ndarray, cell_type: str = "hex", global_connectivity: bool = True, distributed_axis: int = 0):
+
+        self.comm = comm
+        self.x = x
+        self.y = y
+        self.z = z
+        self.ndim = x.ndim
+        self.cell_type = cell_type
+        self.distributed_axis = distributed_axis
+
+        # Outputs
+        self.Points = None
+        self.Connectivity = None
+        self.Offsets = None
+        self.Types = None
+        self.Sizes = None
+        self.NumberOfPoints = None
+        self.NumberOfCells = None
+        self.NumberOfConnectivityIds = None
+        self.shape = None
+        self.data = None
+        
+        if self.distributed_axis != 0:
+            raise NotImplementedError("Only distribution along the first axis (0) is currently implemented for VTK meshes")
+
+        if self.comm.Get_size() > 1:
+            self.global_conn = global_connectivity
+        else:
+            self.global_conn = False # In one rank, the serial connectivity is global
+
+        if self.cell_type != "hex":
+            raise NotImplementedError("Only hexahedral meshes are currently supported for VTK meshes")
+        elif self.cell_type == "hex":
+            if self.ndim == 3:
+                if not self.global_conn:
+                    self.data = set_up_hex_vtk_mesh_local(x, y, z)
+                else:
+                    self.data = set_up_hex_vtk_mesh_global(self.comm, x, y, z, distributed_axis)
+            elif x.ndim == 4:
+                if not self.global_conn:
+                    self.data = set_up_hex_vtk_mesh_sem_local(x, y, z)
+                else:
[…]
