### 2026-03-06 09361bd
Fixed name assignment when closing
