### 2026-03-06 a93f20b
Modified the init docstrings
