### 2026-03-06 e50f92e
Add a catalyst interface
