+++ pysemtools/io/catalyst/insitu_visualization.py
@@ -1,0 +1,173 @@
+""" Module that interfaces with catalyst from a vtk mesh"""
+
+import os
+from pathlib import Path
+import numpy as np
+
+from mpi4py import MPI
+from shapely import node
+import paraview
+import catalyst
+import catalyst_conduit as conduit
+from ...monitoring import Logger
+from ...datatypes.msh_vtk import VTKMesh, VTK_HEXAHEDRON
+
+CATALYST_IMPLEMENTATION_NAME = str(os.getenv("CATALYST_IMPLEMENTATION_NAME", ""))
+CATALYST_IMPLEMENTATION_PATH = str(os.getenv("CATALYST_IMPLEMENTATION_PATH", ""))
+
+class CatalystSession:
+    """
+    Class to interface with catalyst
+        
+    Interface with catalyst to set up a session and load the pipeline.
+
+    Parameters
+    ----------
+    comm : MPI.Comm
+    pipeline : str
+        Path to the saved ParaView Catalyst state file (e.g. "pipeline.py").
+    channel : str
+        Name of the catalyst channel to use (must match the registrationName in your saved state script, e.g. "rbc00001.vtkhdf").
+    implementation_name : str
+        Optional name of the catalyst implementation to use. Also set with env variable CATALYST_IMPLEMENTATION_NAME
+        e.g. export CATALYST_IMPLEMENTATION_NAME=paraview
+    implementation_path : str
+        Optional path to the catalyst implementation. Also set with env variable CATALYST_IMPLEMENTATION_PATH
+        e.g. export CATALYST_IMPLEMENTATION_PATH=/path/to/paraview/lib/catalyst
+    """
+
+    def __init__(self, comm : MPI.Comm, pipeline: str, channel: str, implementation_name = None, implementation_path: str = None):
+
+        self.comm = comm
+        self.pipeline = pipeline
+        self.channel = channel
+        self.implementation_name = implementation_name or CATALYST_IMPLEMENTATION_NAME or ""
+        self.implementation_path = implementation_path or CATALYST_IMPLEMENTATION_PATH or ""
+        self.log = Logger(comm=self.comm, module_name="CatalystSession")
+        self.vtk = None
+        self.init_node = None
+        self.mesh_node = None
+        self.exec_node = None
+
+        if self.implementation_name == "" or self.implementation_path == "":
+            raise ValueError("You must specify the catalyst implementation name and path either via environment variables CATALYST_IMPLEMENTATION_NAME and CATALYST_IMPLEMENTATION_PATH, or by passing them to the constructor.")
+
+        # Create conduit node for catalyst initialization
+        self.init_node = conduit.Node()
+        self.init_node["catalyst/scripts/script/filename"] = str(Path(self.pipeline).resolve())
+        self.init_node["catalyst/mpi_comm"] = comm.py2f()
+
+        # Define implementation
+        self.init_node["catalyst_load/implementation"] = self.implementation_name
+        self.init_node["catalyst_load/search_paths/paraview"] = self.implementation_path
+
+        catalyst.initialize(self.init_node)
+
+        self.log.write("info", f"Initialized CatalystSession - implementation: {self.implementation_name} - path: {self.implementation_path} - pipeline: {self.pipeline}")
+
+    def set_mesh(self, x: np.ndarray, y: np.ndarray, z: np.ndarray, cell_type: str = "hex"):
+        """ Set the mesh for the catalyst session / conduit node
+
+        Parameters
+        ----------
+        x: np.ndarray
+            array with local x coordinates on this rank.
+        y: np.ndarray
+            array with local y coordinates on this rank.
+        z: np.ndarray
+            array with local z coordinates on this rank.
+        cell_type: str
[…]
