### 2026-03-06 5c8f72e
Added docs for catalyst
