+++ docs/_sources/pages/generated/pysemtools.io.catalyst.rst.txt
@@ -1,0 +1,12 @@
+﻿pysemtools.io.catalyst
+======================
+
+.. automodule:: pysemtools.io.catalyst
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      CatalystSession
+   
