+++ docs/_sources/pages/generated/pysemtools.interpolation.rst.txt
@@ -1,0 +1,12 @@
+﻿pysemtools.interpolation
+========================
+
+.. automodule:: pysemtools.interpolation
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Probes
+   
