+++ docs/_sources/pages/generated/pysemtools.comm.rst.txt
@@ -1,0 +1,12 @@
+﻿pysemtools.comm
+===============
+
+.. automodule:: pysemtools.comm
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Router
+   
