+++ docs/_sources/pages/generated/pysemtools.datatypes.rst.txt
@@ -1,0 +1,18 @@
+﻿pysemtools.datatypes
+====================
+
+.. automodule:: pysemtools.datatypes
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Coef
+      Field
+      FieldRegistry
+      Mesh
+      MeshConnectivity
+      MeshPartitioner
+      VTKMesh
+   
