### 2026-03-07 628de9e
Correct C wrapper reference in README
Updated the C wrapper reference from 'CC' to 'mpicc' for clarity.
