+++ 3rd_party/cmake/adios2.cmake
@@ -1,0 +1,101 @@
+include(ExternalProject)
+
+set(ADIOS2_VERSION "2.10.1" CACHE STRING "ADIOS2 version to install")
+set(ADIOS2_GIT_TAG "" CACHE STRING "Explicit ADIOS2 git tag override")
+
+set(BZIP2_VERSION "1.0.8" CACHE STRING "BZip2 version to install")
+set(BZIP2_URL
+    "https://www.sourceware.org/pub/bzip2/bzip2-${BZIP2_VERSION}.tar.gz"
+    CACHE STRING "BZip2 source archive URL")
+
+if(ADIOS2_GIT_TAG STREQUAL "")
+  set(_adios2_git_tag "v${ADIOS2_VERSION}")
+else()
+  set(_adios2_git_tag "${ADIOS2_GIT_TAG}")
+endif()
+
+message(STATUS "ADIOS2 version: ${ADIOS2_VERSION}")
+message(STATUS "ADIOS2 git tag: ${_adios2_git_tag}")
+message(STATUS "BZip2 version: ${BZIP2_VERSION}")
+message(STATUS "BZip2 URL: ${BZIP2_URL}")
+
+set(BZIP2_INSTALL_PREFIX "${THIRD_PARTY_INSTALL_PREFIX}")
+set(BZIP2_INCLUDE_DIR "${BZIP2_INSTALL_PREFIX}/${THIRD_PARTY_INCLUDEDIR}")
+set(BZIP2_LIBRARY_RELEASE "${BZIP2_INSTALL_PREFIX}/${THIRD_PARTY_LIBDIR}/libbz2.so")
+set(BZIP2_LIBRARY_DEBUG   "${BZIP2_INSTALL_PREFIX}/${THIRD_PARTY_LIBDIR}/libbz2.so")
+
+if(ADIOS2_USE_PYTHON)
+  find_package(Python3 REQUIRED COMPONENTS Interpreter)
+  set(_python_args
+    -DADIOS2_USE_Python=ON
+    -DPython_EXECUTABLE=${Python3_EXECUTABLE}
+    -DPYTHON_EXECUTABLE=${Python3_EXECUTABLE}
+    -DPython_FIND_STRATEGY=LOCATION
+  )
+else()
+  set(_python_args
+    -DADIOS2_USE_Python=OFF
+  )
+endif()
+
+if(ADIOS2_USE_FORTRAN)
+  set(_fortran_arg -DADIOS2_USE_Fortran=ON)
+else()
+  set(_fortran_arg -DADIOS2_USE_Fortran=OFF)
+endif()
+
+if(ADIOS2_USE_BZIP2)
+  set(_bzip2_args
+    -DADIOS2_USE_BZip2=ON
+    -DBZIP2_INCLUDE_DIR=${BZIP2_INCLUDE_DIR}
+    -DBZIP2_LIBRARY_RELEASE=${BZIP2_LIBRARY_RELEASE}
+    -DBZIP2_LIBRARY_DEBUG=${BZIP2_LIBRARY_DEBUG}
+  )
+
+  ExternalProject_Add(bzip2_ext
+    URL ${BZIP2_URL}
+    DOWNLOAD_DIR ${THIRD_PARTY_DOWNLOAD_DIR}
+    SOURCE_DIR ${THIRD_PARTY_BUILD_DIR}/src/bzip2-${BZIP2_VERSION}
+    BUILD_IN_SOURCE 1
+    CONFIGURE_COMMAND ""
[…]
