### 2026-03-07 db59aab
adios2 installation works
