+++ 3rd_party/cmake/paraview.cmake
@@ -1,0 +1,107 @@
+include(ExternalProject)
+
+set(PARAVIEW_VERSION "6.0.1" CACHE STRING "ParaView version to install")
+set(PARAVIEW_GIT_TAG "" CACHE STRING "Explicit ParaView git tag override")
+
+if(PARAVIEW_GIT_TAG STREQUAL "")
+  set(_paraview_git_tag "v${PARAVIEW_VERSION}")
+else()
+  set(_paraview_git_tag "${PARAVIEW_GIT_TAG}")
+endif()
+
+if(PARAVIEW_VARIANT STREQUAL "DESKTOP")
+  set(_paraview_use_qt ON)
+elseif(PARAVIEW_VARIANT STREQUAL "SERVER")
+  set(_paraview_use_qt OFF)
+else()
+  message(FATAL_ERROR
+    "Invalid PARAVIEW_VARIANT='${PARAVIEW_VARIANT}'. Use SERVER or DESKTOP.")
+endif()
+
+message(STATUS "ParaView version: ${PARAVIEW_VERSION}")
+message(STATUS "Resolved ParaView git tag: ${_paraview_git_tag}")
+message(STATUS "ParaView variant: ${PARAVIEW_VARIANT}")
+message(STATUS "ParaView Qt GUI enabled: ${_paraview_use_qt}")
+
+set(_paraview_args
+  -DCMAKE_INSTALL_PREFIX=${THIRD_PARTY_INSTALL_PREFIX}
+  -DCMAKE_INSTALL_BINDIR=${THIRD_PARTY_BINDIR}
+  -DCMAKE_INSTALL_INCLUDEDIR=${THIRD_PARTY_INCLUDEDIR}
+  -DCMAKE_INSTALL_LIBDIR=${THIRD_PARTY_LIBDIR}
+
+  -DCMAKE_C_COMPILER=${CMAKE_C_COMPILER}
+  -DCMAKE_CXX_COMPILER=${CMAKE_CXX_COMPILER}
+
+  -DPARAVIEW_USE_PYTHON=${PARAVIEW_USE_PYTHON}
+  -DPARAVIEW_USE_MPI=${PARAVIEW_USE_MPI}  
+  -DVTK_SMP_IMPLEMENTATION_TYPE=${PARAVIEW_SMP_IMPLEMENTATION_TYPE}
+  -DCMAKE_BUILD_TYPE=Release
+  -DPARAVIEW_ENABLE_CATALYST=${PARAVIEW_ENABLE_CATALYST} 
+  -DPARAVIEW_USE_QT=${_paraview_use_qt}
+)
+
+# Python forwarding
+if(PARAVIEW_USE_PYTHON)
+  if(PYTHON_EXECUTABLE)
+    if(NOT EXISTS "${PYTHON_EXECUTABLE}")
+      message(FATAL_ERROR "PYTHON_EXECUTABLE does not exist: ${PYTHON_EXECUTABLE}")
+    endif()
+
+    list(APPEND _paraview_args
+      -DPython3_EXECUTABLE=${PYTHON_EXECUTABLE}
+      -DPython_EXECUTABLE=${PYTHON_EXECUTABLE}
+      -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE}
+      -DPython3_FIND_STRATEGY=LOCATION
+      -DPython_FIND_STRATEGY=LOCATION
+    )
+
+    if(Python3_ROOT_DIR)
+      list(APPEND _paraview_args
+        -DPython3_ROOT_DIR=${Python3_ROOT_DIR}
+      )
+    endif()
+  else()
+    find_package(Python3 REQUIRED COMPONENTS Interpreter)
+    list(APPEND _paraview_args
[…]
