### 2026-03-09 5b27ae5
Added mpi4py to the available options and updated readme
