### 2026-03-09 511432a
Added known instructions
