### 2026-03-09 cbeacba
Set 1.3.0 version
