+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "extremeflow-pysemtools"
-version = "1.3.0.dev0"
+version = "1.4.0.dev0"
 readme = "README.md"
 requires-python = ">= 3.10"
 dependencies = [
