### 2026-03-09 #8767 @adperezm
Hi @finsberg,

Here is the info:
Version: 1.3.0
Zenodo: https://zenodo.org/records/18926958
doi: https://doi.org/10.5281/zenodo.18926958

Let me know if this information is enough.
