### 2026-03-09 #8767 @finsberg
Thanks @adperezm , this is enough information. Could you also make sure that the affiliations in the Zenodo archive matches the affiliations in the paper?
