### 2026-03-10 #8767 @adperezm
Thank you for the feedback @finsberg. They match now. I also removed one space in one of the the affiliations in the paper (FLOW Dept. Engineering Mechanics, KTH Royal Institute of Technology). Just saying in case we should generate a new proof.
