### 2026-03-10 #8767 @adperezm
The proof looks fine from my side.
