+++ 3rd_party/CMakeLists.txt
@@ -80,6 +80,33 @@ set(PYTHON_EXECUTABLE ""
 set(Python3_ROOT_DIR ""
     CACHE PATH "Optional root dir for the selected Python installation")
 
+
+# =========== This block just finds the right python version and sets a relative install path, since adios2.10.1 was misbihaving
+# =========== remove this if newer versions of adios2 behave like the other libraries
+# Get the exact name of where the python bindings should be installed for ADIOS2.
+## Set the python executable
+if(PYTHON_EXECUTABLE)
+  if(NOT EXISTS "${PYTHON_EXECUTABLE}")
+    message(FATAL_ERROR "PYTHON_EXECUTABLE does not exist: ${PYTHON_EXECUTABLE}")
+  endif()
+  set(_third_party_python_exe "${PYTHON_EXECUTABLE}")
+else()
+  find_package(Python3 REQUIRED COMPONENTS Interpreter)
+  set(_third_party_python_exe "${Python3_EXECUTABLE}")
+endif()
+## Execute python to see obtain its version
+execute_process(
+  COMMAND "${_third_party_python_exe}" -c
+          "import sys; print(f'{sys.version_info.major}.{sys.version_info.minor}')"
+  OUTPUT_VARIABLE THIRD_PARTY_PYTHON_VERSION
+  OUTPUT_STRIP_TRAILING_WHITESPACE
+)
+## Set the variable that adios2 can use
+set(THIRD_PARTY_PYTHONDIR
+    "${THIRD_PARTY_LIBDIR}/python${THIRD_PARTY_PYTHON_VERSION}/site-packages"
+    CACHE STRING "Relative install dir for Python packages under the third-party prefix")
+#===============================================================
+
 file(MAKE_DIRECTORY "${THIRD_PARTY_INSTALL_PREFIX}")
 file(MAKE_DIRECTORY "${THIRD_PARTY_DOWNLOAD_DIR}")
 file(MAKE_DIRECTORY "${THIRD_PARTY_BUILD_DIR}")
