+++ 3rd_party/env.sh.in
@@ -1,0 +1,15 @@
+#!/usr/bin/env bash
+
+_this_script="${BASH_SOURCE[0]}"
+_third_party_prefix="$(cd "$(dirname "$_this_script")" && pwd)"
+
+export THIRD_PARTY_PREFIX="$_third_party_prefix"
+export PATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_BINDIR@:$PATH"
+export LD_LIBRARY_PATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_LIBDIR@:$THIRD_PARTY_PREFIX/lib64:${LD_LIBRARY_PATH:-}"
+export LIBRARY_PATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_LIBDIR@:$THIRD_PARTY_PREFIX/lib64:${LIBRARY_PATH:-}"
+export CPATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_INCLUDEDIR@:${CPATH:-}"
+export CMAKE_PREFIX_PATH="$THIRD_PARTY_PREFIX:${CMAKE_PREFIX_PATH:-}"
+export PKG_CONFIG_PATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_LIBDIR@/pkgconfig:$THIRD_PARTY_PREFIX/lib64/pkgconfig:$THIRD_PARTY_PREFIX/share/pkgconfig:${PKG_CONFIG_PATH:-}"
+
+# Python packages installed under the local prefix
+export PYTHONPATH="$THIRD_PARTY_PREFIX/@THIRD_PARTY_PYTHONDIR@:${PYTHONPATH:-}"
