### 2026-03-11 8ee0823
Modified catalyst path for paraview inst
